% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_tr.R
\name{acceleration}
\alias{acceleration}
\title{Calculate acceleration.}
\usage{
acceleration(x, y, f)
}
\arguments{
\item{x}{X coordinates.}

\item{y}{Y coordinates.}

\item{f}{Always use 'f' or 'frame' (equivalent). Used to order by frame
before calculation.}
}
\value{
A vector of accelerations in px per frame^2.
}
\description{
Calculate acceleration.
}

