% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_tr.R
\name{add_acceleration}
\alias{add_acceleration}
\title{Add acceleration to tracks object.}
\usage{
add_acceleration(tracks)
}
\arguments{
\item{tracks}{A tracks object.}
}
\value{
tracks object
}
\description{
Add acceleration to tracks object.
}

