% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_soc.R
\name{add_nip_dist}
\alias{add_nip_dist}
\title{Add pairwise distance based on tip to ellipse.}
\usage{
add_nip_dist(tracks, n = 20)
}
\arguments{
\item{tracks}{A tracks object.}

\item{n}{How many}
}
\value{
A tracks object
}
\description{
Add the distance between the pair of animals for each entry in the $soc
table. This distance is from the mouth of fish 1 (tip of the ellipse) to
an approximately closest point on the ellipse for fish 2.
}
\details{
It does not get the closest point at the moment, as this calculation is quite
complicated, but does a numerical approximation based on n points along the
ellipse.
}

