% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_soc.R
\name{add_nip_dist_velocity}
\alias{add_nip_dist_velocity}
\title{Add change in pairwise distance based on tip to ellipse.}
\usage{
add_nip_dist_velocity(tracks)
}
\arguments{
\item{tracks}{A tracks object.}
}
\value{
A tracks object.
}
\description{
Add change in pairwise distance based on tip to ellipse.
}
\seealso{
add_nip_dist, add_pair_dist_velocity
}

