% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_soc.R
\name{add_pair_dist}
\alias{add_pair_dist}
\title{Add pairwise distance.}
\usage{
add_pair_dist(tracks)
}
\arguments{
\item{tracks}{A tracks object.}
}
\value{
A tracks object
}
\description{
Add the distance between centroids for the pair of animals for each entry in
the $soc table.
}

