% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_pairs.R
\name{add_pair_dist_acceleration}
\alias{add_pair_dist_acceleration}
\title{Add acceleration in distance between pairs.}
\usage{
add_pair_dist_acceleration(tracks)
}
\arguments{
\item{tracks}{A tracks object.}
}
\value{
A tracks object.
}
\description{
Add the change in the change in distance between the pair of animals for each
entry in the $pairs table. Positive values mean the animals are increasing
the speed at which their distance in changing.
}

