% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_pairs.R
\name{add_pair_dist_velocity}
\alias{add_pair_dist_velocity}
\title{Add change in distance between pairs}
\usage{
add_pair_dist_velocity(tracks)
}
\arguments{
\item{tracks}{A tracks object.}
}
\value{
A tracks object.
}
\description{
Add the change in distance between the pair of animals for each entry in the
$pairs table. Negative values mean the animals are moving closer to each
other, postitive values means they are moving apart.
}

