% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_positions.R
\name{add_relative_positions}
\alias{add_relative_positions}
\title{Add the relative positions of each pair}
\usage{
add_relative_positions(tracks)
}
\arguments{
\item{tracks}{A tracks object.}
}
\value{
A data.frame with xbin, ybin and n (count).
}
\description{
Add the relative positions of each pair
}
