% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_tr.R
\name{add_speed}
\alias{add_speed}
\title{Add speed to tracks object}
\usage{
add_speed(tracks)
}
\arguments{
\item{tracks}{A tracks object.}
}
\value{
tracks object
}
\description{
Add speed to tracks object
}
\examples{
add_speed(guppies)
}

