% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_tr.R
\name{add_to_tr}
\alias{add_to_tr}
\title{Internal function to add a new calculated variable calculation to a tracks
object.}
\usage{
add_to_tr(tracks, .calc, ...)
}
\arguments{
\item{tracks}{A tracks object.}

\item{.calc}{A function of the form function(tracks, ...), which returns a
a data frame. Should return tr plus any new columns necessary.}

\item{...}{Any extra arguments to supply to .calc.}
}
\value{
tracks object
}
\description{
This function will take .calc and apply it to the tr portion of the tracks
object, feeding it portions of tr grouped by trial and animal. It will then
summarize this new variable to the other parts of the tracks object using
.summary (typically this would be mean, or median etc.).
}
\section{Important Note}{

 .calc should return something of the form:
     \code{dplyr::tbl_df(data.frame(d, var = x))} and that name var should be
 passed to this function as the first argument in .names.
}

