% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_tr.R
\name{add_turn}
\alias{add_turn}
\title{Add turning angle to tracks object}
\usage{
add_turn(tracks)
}
\arguments{
\item{tracks}{A tracks object.}
}
\value{
tracks object
}
\description{
This adds the turning angle of the path, that is, based on coordinates, not
orientation.
}
\examples{
add_turn(guppies)
}

