% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjFeatureSet.R
\name{as.list.FeatureSet}
\alias{as.list.FeatureSet}
\alias{as.data.frame.ObjFeatureSet}
\title{Export a ObjFeatureSet as list}
\usage{
\method{as.list}{FeatureSet}(x, ...)

\method{as.data.frame}{ObjFeatureSet}(x, row.names = NULL, optional = FALSE,
  ..., long = FALSE)
}
\arguments{
\item{x}{A ObjFeatureSet.}

\item{...}{Other named arguments (currently unused).}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed. Only applicable in conjunction with long = FALSE.}

\item{optional}{Currently unused in this implementation.}

\item{long}{A boolean indicating whether the data.frame should be returned in 'long' (rather than 'wide') format.}
}
\value{
A list.

A data.frame.
}
