% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgrounder.R
\name{backgrounder}
\alias{backgrounder}
\title{Reconstruct the Background of a Video}
\usage{
backgrounder(
  video,
  n = 10,
  method = "median",
  prob = 0.025,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{video}{A video object as produced by \code{\link[Rvision]{video}}.}

\item{n}{The number of images of the video that will be used to reconstruct
the background.}

\item{method}{The name of a method to reconstruct the background. There are
currently 4 methods available: "median" (uses the median value of each pixel
as background), "min" (uses the minimum value of each pixel as background),
"max" (uses the maximum value of each pixel as background), and "quant"
(uses an arbitrary quantile value of each pixel as background).}

\item{prob}{If \code{method = "quant"}, the quantile value to use.}

\item{start, end}{The start and end frames of the video to use for the
background reconstruction. If not set, the first and last frames will be
used.}
}
\value{
And \code{\link[Rvision]{Image}} object.
}
\description{
This function attempts to reconstruct the background of a video
 that was filmed from a fixed point of view.
}
\examples{
cctv <- Rvision::video(system.file("sample_vid/Walk3.mp4", package = "Rvision"))
background <- backgrounder(cctv, 20)
plot(background)

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
