% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_positions.R
\name{bin_relative_positions}
\alias{bin_relative_positions}
\title{Count the relative positions of each pair binned on X and Y}
\usage{
bin_relative_positions(tracks, nbins = 100)
}
\arguments{
\item{tracks}{A tracks object.}

\item{nbins}{The number of bins you want. The same size bins are used for X and Y.}
}
\value{
A data.frame with xbin, ybin and n (count).
}
\description{
Count the relative positions of each pair binned on X and Y
}
