% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_trial_operation.R
\name{by_trial_operation}
\alias{by_trial_operation}
\title{Perform an arbitrary operation with a different constant for each trial.}
\usage{
by_trial_operation(tracks, tab, var, operation = `+`, trial = "trial")
}
\arguments{
\item{trial}{}
}
\description{
Perform an arbitrary operation with a different constant for each trial.
}
