% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_pairs.R
\name{calc_speed_lag}
\alias{calc_speed_lag}
\title{Calculate pairwise speed correlation lag.}
\usage{
calc_speed_lag(tracks, range = 100, time_bin = NULL)
}
\arguments{
\item{tracks}{A tracks object.}

\item{range}{The range of frames for which lags should be calculated, e.g.
100 means from lags from -100 to +100 are evaluated.}
}
\value{
A data.frame.
}
\description{
Pairwise correlations can be used to determine leader follower relationships.
Nagy et al. Nature (464) 2010 used directional correlations, this function
allows for speed correlations.
}
\details{
This function returns two variables per animal pair per trial;
\code{lag} is the delay where the correlation was maximal (positive means),
and \code{cor} is the magnitude of the correlation.
}

