% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDist.r
\name{calcDist}
\alias{calcDist}
\title{Compute euclidian distance between 2 points}
\usage{
calcDist(data, spaceRes = 1)
}
\arguments{
\item{data}{a dataframe with 2 lines and 2 or 3 columns (x,y,(z))}
}
\description{
Compute euclidian distance between 2 points
}
\examples{
data2D <- data.frame(x =c(0,1), y=c(0,1))
calcDist(data2D)
data3D <- data.frame(x =c(0,1), y=c(0,1), z=c(0,1))
calcDist(data3D)
}
\author{
JuG
}
