% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMSD.r
\name{calcMSD}
\alias{calcMSD}
\title{Calculate Mean Square Distance}
\usage{
calcMSD(data, N)
}
\arguments{
\item{data}{data frame (including trace, x, y and t)}

\item{N}{for MSD, dt should be up to 1/N of number of data points (4 recommended)}
}
\description{
Calculate Mean Square Distance
}
\examples{
xmlPath <- "//Users/jgodet/Seafile/MaBibliotheque/Hanna/1_3_MMStack_Pos0.ome.xml"
xmlPath <- "/Users/jgodet/Seafile/MaBibliotheque/Code/TrackMate/nmeth.2808-sv1.xml"
data <- readTrackMateXML(XMLpath = xmlPath)
data$jump<-jump(data, spaceRes=1)
msdMat <- calcMSD(data, N=3)
plotMSD(msdMat, col='blue')
}
\author{
JuG
}
