% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_tr.R
\name{change}
\alias{change}
\title{Safely calculate the change over time for a variable.}
\usage{
change(x, frame, n = 1)
}
\arguments{
\item{x}{A vector.}

\item{n}{Time lag in frames.}
}
\value{
A vector.
}
\description{
Use this convenience function within calls to mutate_.tracks.
}

