% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_object.R
\name{check_complete}
\alias{check_complete}
\title{Report track completeness per trial}
\usage{
check_complete(tracks, vars = c("X", "Y"), lower_limit = 95,
  parallel = TRUE)
}
\arguments{
\item{tracks}{A tracks object}

\item{vars}{The variables in the $tr table that need to be present to count
towards completeness.}

\item{lower_limit}{If a percentage is below this limit, the cell in the table
will be marked with a *.}
}
\value{
Prints a table, silently returns the table too.
}
\description{
The report gives two statistics:
- complete: which is simply the percentage of frame by animal by trial
  combinations for which the vars are there.
- per frame: which checks for what percentage of frames all animals are
complete. This is useful for social parameters, for which all individuals are
needed.
}
\details{
If some frames were poorly tracked, the two parameters will be similar. If
there are large differences, then there is many frames in which a portion of
the individuals gets.
}
\section{Note}{

This is only useful if their are persistent id's across the video. If there
are any spurious detections, this will lead to very low scores.
}

