% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_row_diff}
\alias{check_row_diff}
\title{Function to check which variables differ across two rows}
\usage{
check_row_diff(tibble_data, k = 1)
}
\arguments{
\item{k}{}
}
\description{
Function to check which variables differ across two rows
}
