% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{checkPandoc}
\alias{checkPandoc}
\title{check for availability of pandoc}
\usage{
checkPandoc()
}
\description{
This function is exported because it is used as a default value
end users should not need to call it directly
}
