% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTraces.r
\name{clusterTraces}
\alias{clusterTraces}
\title{cluster traces in space (and/or time) using DBscan}
\usage{
clusterTraces(dtf, coord = c("x", "y"), eps, minPts, method, ...)
}
\arguments{
\item{dtf}{data frame with coordinates and traces ID}

\item{coord}{coordinates dataframe (x,y) or (x,y,t)}

\item{eps}{Reachability distance, see Ester et al. (1996).}

\item{minPts}{Reachability minimum no. of points, see Ester et al. (1996).}

\item{method}{"dist" treats data as distance matrix (relatively fast but memory expensive), "raw" treats data as raw data and avoids calculating a distance matrix (saves memory but may be slow), "hybrid" expects also raw data, but calculates partial distance matrices (very fast with moderate memory requirements).}
}
\value{

}
\description{
cluster traces in space (and/or time) using DBscan
}
\examples{
xmlPath <- "//Users/jgodet/Seafile/MaBibliotheque/Hanna/1_3_MMStack_Pos0.ome.xml"
data <- readTrackMateXML(XMLpath = xmlPath)
data$jump<-jump(data, spaceRes=1)
bacteria <- clusterTraces(dtf = data, eps = .2,minPts = 100)
table(bacteria)
bact1 <- data[which(bacteria==7),]
summary(bact1)
bact1 \%>\% select(x,y) \%>\% plot(., asp=1)
drawRod(data = getContour(bact1),col='green')
drawRod(data = getContour(cleanNearest(data = bact1,k = 3)),col='blue')
}
\author{
JuG
}
