% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjFeatureSet.R
\name{as.list.ScriptNodeInfo}
\alias{as.list.ScriptNodeInfo}
\title{Export a ScriptNodeInfo object as a list}
\usage{
\method{as.list}{ScriptNodeInfo}(x, ...)
}
\arguments{
\item{x}{A ScriptNodeInfo object.}

\item{...}{Other named arguments (currently unused).}
}
\value{
A list.
}
