% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-dplyr_verbs.R
\name{collapse_identities}
\alias{collapse_identities}
\title{Collapse multiple identies in fewer.}
\usage{
collapse_identities(tracks, n = 1)
}
\arguments{
\item{tracks}{A \code{tracks} object.}

\item{n}{The number of animals. Currently only 1 is supported.}
}
\value{
A \code{tracks} object.
}
\description{
This function allows one to assign several duplicated identities over time
to a fewer number of actual animals. When more identities exist within a
frame than there is animals, data is filtered out.
}

