% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjFeatureSet.R
\name{collapse_to_str}
\alias{collapse_to_str}
\title{Collapse a list to a comma-delimited character vector}
\usage{
collapse_to_str(l, with.names = FALSE)
}
\arguments{
\item{l}{A list}

\item{with.names}{A boolean indicating whether the list is named. False by default.}
}
\value{
A comma-delimited character vector
}

