% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sources.R
\name{combine_sources}
\alias{combine_sources}
\title{Integrate data from multiple tracking sources}
\usage{
combine_sources(tracks1, tracks2, err_cutoff = 25^2, report = TRUE)
}
\arguments{
\item{tracks1}{Tracks object to be combined.}

\item{tracks2}{The other tracks object to be combined.}

\item{err_cutoff}{Hard limit for error (squared distance), that is deemed
unreliable as a match. The default is 25 ^ 2, which means the centroids
estimated are allowed to be 25 pixels apart.}

\item{report}{Whether a small report should be printed to the console.}
}
\value{
A tracks object.
}
\description{
This function will attempt to combine data from trials that have been tracked
by multiple trackers ('sources'). This is useful if one want to leverage the
individual strengths of multiple trackers (e.g. combine orientation and
ellipse data from Ctrax with resolved identities from idTracker). It can also
be used to exert quality control, by excluding observations that trackers
don't agree on.
}
\details{
All variables that are supplied by both trackers will be used to match
identities between sources. The error will be calculated and added to the $tr
table, along with the mean values for the two sources. All other variables
will be maintained. In some cases, precedence may be given when a source
clearly is better suited for that variable (e.g. when combining idTracker
with Ctrax, identities will always be based on idTracker.)

The methods tend to be designed for reasonable speeeds, and may use a lot
of memory.

Animal names _will_ be lost.
}
\section{}{
 Note on parallel performance:
  Some of the current implementation limitations fo multidplyr force us to
  start and close several clusters during combining.
}

