% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_nas}
\alias{count_nas}
\title{Count the number of NAs in each variable}
\usage{
count_nas(data, sort = FALSE, return_count = FALSE)
}
\arguments{
\item{data}{dataset}

\item{sort}{logical. If TRUE, the final count will be sorted by the proportion of NAs}

\item{return_count}{logical. If FALSE, the count is printed, and the dataset is invisibly returned. If TRUE, the count itself is returned.}
}
\description{
Count the number of NAs in each variable
}
