% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_print}
\alias{count_print}
\title{Count in middle of a pipeline and print}
\usage{
count_print(data, ...)
}
\arguments{
\item{data}{dataset}

\item{...}{other arguments to pass from count}
}
\value{

}
\description{
Count in middle of a pipeline and print
}
