% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_prop}
\alias{count_prop}
\title{Create a relative frequency table and print}
\usage{
count_prop(data, ..., return_count = FALSE)
}
\arguments{
\item{data}{dataset}

\item{...}{other arguments to pass to}

\item{return_count}{logical. If FALSE, the count is printed, and the dataset is invisibly returned. If TRUE, the count itself is returned.}
}
\description{
Create a relative frequency table and print
}
