% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpheroCylinderCoord.r
\name{createSpheroCylinderCoord}
\alias{createSpheroCylinderCoord}
\title{Do something}
\usage{
createSpheroCylinderCoord(
  N = 1000,
  l = 2,
  r = 0.3,
  initialPos = c(1, 1, 0),
  rot = 0
)
}
\arguments{
\item{N}{umber of points}

\item{l}{length of the rod}

\item{r}{radius}

\item{initialPos}{Initial position}

\item{rot}{Rotation}
}
\value{

}
\description{
None
}
\details{
None
}
\examples{
coord <- createSpheroCylinderCoord(rot=10)
plot(coord,asp=1)

}
\author{
JuG
}
