% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestEllipse.r
\name{createTestEllipse}
\alias{createTestEllipse}
\title{Create a test ellipse}
\usage{
createTestEllipse(Rx = 300, Ry = 200, Cx = 250, Cy = 150,
  Rotation = 0.4, N = 200, NoiseLevel = 10)
}
\arguments{
\item{Rx}{X-radius}

\item{Ry}{Y-radius}

\item{Cx}{X-center}

\item{Cy}{Y-center}

\item{Rotation}{Radians}

\item{N}{Nb of points}

\item{NoiseLevel}{Gaussian Noise level}
}
\description{
Create a test ellipse
}
\examples{


}
\author{
JuG
}
