% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{generateTags}
\alias{generateTags}
\title{customizing-metadata}
\usage{
generateTags(object)
}
\arguments{
\item{object}{The object to generate tags for}
}
\value{
A character vector of tags to associate with \code{object} during the
recording process
}
