% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R,
%   R/accessor-methods.R
\docType{methods}
\name{dataLabels<-}
\alias{dataLabels<-}
\alias{dataLabels}
\alias{dataLabels,PlotFeatureSet-method}
\alias{dataLabels,ggplot-method}
\alias{dataLabels,trellis-method}
\alias{dataLabels,gTree-method}
\alias{dataLabels<-,PlotFeatureSet-method}
\alias{dataLabels,PlotFeatureSet-method}
\title{Mutator method for PlotFeatureSet "varlabels" slot.}
\usage{
dataLabels(object) <- value

dataLabels(object)

\S4method{dataLabels}{PlotFeatureSet}(object)

\S4method{dataLabels}{ggplot}(object)

\S4method{dataLabels}{trellis}(object)

\S4method{dataLabels}{gTree}(object)

\S4method{dataLabels}{PlotFeatureSet}(object) <- value

\S4method{dataLabels}{PlotFeatureSet}(object)
}
\arguments{
\item{object}{An object of (super)class PlotFeatureSet. For accession, also plot objects of class ggplot or trellis.}

\item{value}{A named list of variable labels of the form list(x = "X axis label", y = "Y axis label", ...).}
}
\value{
A modified object of (super)class PlotFeatureSet.

A named list of variable labels of the form list(x = "X axis label", y = "Y axis label", ...).
}
