% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{dataNames}
\alias{dataNames}
\title{Accessor method for the variable names of the plot object. Not an exported method.}
\usage{
dataNames(object)
}
\arguments{
\item{object}{An object of (super)class PlotFeatureSet. For accession, also plot objects of class ggplot or trellis.}
}
\value{
A named list of variable names of the form list(x = "x.name", y = "y.name", groups = list(...)).
}
\description{
Get the variable names from a PlotFeatureSet
}
