% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotrack.R
\name{defaultVTDB}
\alias{defaultVTDB}
\title{default VTDB}
\usage{
defaultVTDB(vtdb)
}
\arguments{
\item{vtdb}{ViztrackrDB. The ViztrackrDB to which objects will be recorded
by default. If missing, the current default is returned.}
}

