% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R
\docType{methods}
\name{describePackage}
\alias{describePackage}
\alias{describePackage,ObjFeatureSet-method}
\title{Accessor method for ObjFeatureSet "analysispkg" slot.}
\usage{
describePackage(object)

\S4method{describePackage}{ObjFeatureSet}(object)
}
\arguments{
\item{object}{An object of (super)class ObjFeatureSet.}
}
\value{
A list holding the name, title, and description of any packages with description files on the current path.
}

