% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_tr.R
\name{distance}
\alias{distance}
\title{Calculate distance.}
\usage{
distance(x = X, y = Y, order_by = frame)
}
\arguments{
\item{x}{X coordinates.}

\item{y}{Y coordinates.}

\item{f}{Always use 'f' or 'frame' (equivalent). Used to order by frame
before calculation.}
}
\value{
A vector of distances in px.
}
\description{
Calculate distance.
}

