% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{DocCollectionRef}
\alias{DocCollectionRef}
\alias{.docrefclass}
\alias{ListBackend-class}
\alias{listbackend}
\alias{ListBackend}
\title{Reference to a DocCollection}
\usage{
ListBackend(lst = list())
}
\description{
A reference class which carries around a DocCollection, suitable
for use as a trackr backend.
}

