% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawRod.r
\name{drawRod}
\alias{drawRod}
\title{Do something}
\usage{
drawRod(data, ...)
}
\arguments{
\item{data}{coordinates of the contour}
}
\description{
Do something
}
\examples{
coord <- createSpheroCylinderCoord( l =1.8,r=.3,rot=120)
plot(coord[,1:2],asp=1,pch='.')
gcont <- getContour(coord[,1:2],drawContour = TRUE)
drawRod(gcont,lty=3)
}
\author{
JuG
}
