% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_track.R
\name{drop_na_track}
\alias{drop_na_track}
\title{drop_na and see how many you drop}
\usage{
drop_na_track(data, ...)
}
\arguments{
\item{...}{arguments to pass to drop_na}
}
\value{

}
\description{
drop_na and see how many you drop
}
