% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ellipse}
\alias{ellipse}
\title{Points on an Ellipse}
\usage{
ellipse(x, y, width, height, angle, npoints = 100)
}
\arguments{
\item{x, y}{Numeric values corresponding to the coordinates of the center of
the ellipse.}

\item{width, height}{Numeric values corresponding to the width and height of
the ellipse.}

\item{angle}{Numeric value corresponding to the angle of the ellipse relative
to the y axis.}

\item{npoints}{The number of points to compute.}
}
\value{
A matrix with two columns corresponding to the x and y coordinates of
 the points.
}
\description{
This functions computes \code{npoints} regularly spaced along an
 ellipse.
}
\examples{
plot(ellipse(0, 0, 30, 50, 33), asp = 1)

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
