% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R
\docType{methods}
\name{saveEnrichedPlot}
\alias{loadEnrichedPlot}
\alias{saveEnrichedPlot}
\alias{saveEnrichedPlot,PlotFeatureSet-method}
\title{Save the plot object owned by an object of class/superclass PlotFeatureSet as an image with metadata.}
\usage{
saveEnrichedPlot(object, filename, ...)

\S4method{saveEnrichedPlot}{PlotFeatureSet}(object, filename, type = "png",
  ...)

loadEnrichedPlot(filename, type = "png", ...)
}
\arguments{
\item{object}{An object of (super)class PlotFeatureSet.}

\item{filename}{The full filename, including path, where the plot should be saved.}

\item{...}{Other named arguments passed on to the class-specific basic image save method.}

\item{type}{The type of metadata-enriched image to save. Default is "png"}
}
\section{Functions}{
\itemize{
\item \code{loadEnrichedPlot}: Load the (R) plot object previously saved as metadata in an image file.
}}

