% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_object.R
\name{expand_tracks}
\alias{expand_tracks}
\title{Add missing components to a tracks object.}
\usage{
expand_tracks(tracks, group = TRUE, pairs = TRUE, animal = TRUE,
  trial = TRUE)
}
\arguments{
\item{tracks}{A tracks object.}

\item{group}{If TRUE, group will be checked for and added if missing.}

\item{pairs}{If TRUE, pairs will be checked for and added if missing.}

\item{animal}{If TRUE, animal will be checked for and added if missing.}

\item{trial}{If TRUE, trial will be checked for and added if missing.}
}
\value{
A tracks object.
}
\description{
This function will add the group, pairs and trial components to a track
object, if they are currently not included. Note that the pairs component
will be increase the size by a lot, if there is more than a few animals per
trial.
}
\details{
Note that this function is similar to running \code{as_tracks(tracks$tr, ...)
}, but safer (i.e. no loss of data).

If group, pairs and trial are all false, it will return tracks.
}
\examples{
expand_tracks(guppies)
}

