% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_results_to_excel}
\alias{export_results_to_excel}
\title{Export progress results to Excel}
\usage{
export_results_to_excel(res_list, file_path)
}
\arguments{
\item{res_list}{output of `track_progress()`}

\item{file_path}{path to save file}
}
\description{
Export progress results to Excel
}
