% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{filter}
\alias{filter}
\alias{filter_}
\alias{filter_.tracks}
\title{Return a subset of a tracks object with matching conditions.}
\usage{
\method{filter_}{tracks}(.data, ..., drop = TRUE, .dots,
  repartition = FALSE)
}
\arguments{
\item{.data}{A tracks object.}

\item{...}{The filtering conditions (logical statements).}

\item{drop}{Whether to drop conflicting aggregate data. If the tracks object
contains data that was aggregated over a variable that is now being used to
filter, drop must be TRUE or an error will be raised. That is, any table
that can not be filtered (i.e. it does not contain the variable used to
filter) will be dropped. Have to opt in, since aggregation may have been
expensive.}

\item{repartition}{If TRUE, will collect and reassign the \code{tr} and
\code{soc} tables. Useful if worker nodes become empty or unbalanced.}
}
\value{
The subsetted tracks object.
}
\description{
Method for the dplyr verb \code{filter}. Filters the tracks object according
to logical filtering statements, in a strict sense. If one of the filtering
conditions cannot be be applied to all tables present, it will either give an
error (\code{drop = FALSE}) or will delete the tables that can't be filtered
from the tracks object (\code{drop = FALSE}), see details. For another
strategy, see \code{find_sections} and \code{summarize_sections}.
}
\section{Strategy on conflicting dependecies}{


  It will apply the fitering to all the applicable sections of the tracks
  object. It will also attempt to check for any dependency problems. This is
  crucial in order to maintain internal consistency in the tracks object. For
  example, if you select a sequence of frames, any data that was aggregated
  over time (such as in the \code{trial} table) will now no longer match. If
  conflicts are found, by default, conflicted tables will be deleted from
  the tracks object, with a message. By setting \code{drop = FALSE}, you can
  have the function error instead.
}

\seealso{
\link[dplyr]{filter}, \link{find_sections},
  \link{summarise_sections}
}
