% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_track.R
\name{filter_track}
\alias{filter_track}
\title{Filter and show count}
\usage{
filter_track(data, ...)
}
\arguments{
\item{data}{dataset}

\item{...}{other arguments to pass to filter}
}
\description{
Filter and show count
}
