% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry.R
\name{find_smallest_enclosing_circle}
\alias{find_smallest_enclosing_circle}
\title{Find smalles enclosing circle.}
\usage{
find_smallest_enclosing_circle(S)
}
\arguments{
\item{S}{A matrix or data.frame of point coordinates with the first column
being x and the second being y.}
}
\value{
A vector of lenght three, center x and y, and radius.
}
\description{
Algorithm by Sven Skyum, from http://dx.doi.org/10.7146/dpb.v19i314.6704.
}

