% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry.R
\name{find_smallest_enclosing_rect}
\alias{find_smallest_enclosing_rect}
\title{Find smalles enclosing rectangle.}
\usage{
find_smallest_enclosing_rect(points)
}
\arguments{
\item{points}{A matrix or data.frame of point coordinates with the first
column being x and the second being y.}
}
\value{
A vector of lenght three, center x and y, and radius.
}
\description{
Algorithm from William Huber (http://gis.stackexchange.com/a/22934) and
}

