% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-dplyr_verbs.R
\name{find_track_sections}
\alias{find_track_sections}
\alias{find_track_sections_}
\title{Retrieve the timestamps for track section based on conditions.}
\usage{
find_track_sections(tracks, ..., tol = 0)

find_track_sections_(tracks, ..., tol = 1, .dots)
}
\arguments{
\item{tracks}{A tracks object.}

\item{...}{Conditions.}

\item{tol}{Combine sequences that are \code{tol} frames apart.}
}
\value{
A tbl_df.
}
\description{
Similar to \code{filter}, but returns a table of sequences grouped by trial,
with a \code{start}, \code{stop} and \code{length} column. It supports a
tolerance level which allows for combining sequences that are close together
(the default of 1 is zero tolerance).
}
\details{
Any variables used as conditions will be looked up in the \code{tr},
\code{pairs} and \code{group} tables and applied when present. Non-existing
variables will not produce an error.

Seperate conditions on different variables with a \code{,}, not \code{&}.
Combine several conditions on the same variable with \code{&}.

Example use case: If one wants to select all sequences of frames where
animals are chasing each other (to plot them, for example), one could filter
for a high mean speed of the pair, and a small pairwise distance.
}

