% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{first_non_na}
\alias{first_non_na}
\title{First non-NA in a vector
Calculate the first value in a vector that's not an NA, otherwise return NA}
\usage{
first_non_na(x)
}
\arguments{
\item{x}{vector}
}
\description{
First non-NA in a vector
Calculate the first value in a vector that's not an NA, otherwise return NA
}
