% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEllipse.r
\name{fitEllipse}
\alias{fitEllipse}
\title{Adjust an ellipse on a contour plot}
\usage{
fitEllipse(x, y = NULL)
}
\arguments{
\item{x}{x.coordinates}

\item{y}{y.coordinates}
}
\description{
Least squares fitting of an ellipse to point data
}
\details{
# Least squares fitting of an ellipse to point data using the algorithm described in Radim Halir & Jan Flusser. 1998
Adapted from the original Matlab code by Michael Bedward (2010)
michael.bedward@gmail.com
Subsequently improved by John Minter (2012)
http://lastresortsoftware.blogspot.fr/2012/09/fitting-ellipse-to-point-data.html
}
\examples{
ellipse <- createTestEllipse(NoiseLevel = 50)
plot(ellipse)
efit <- fitEllipse(ellipse)
e <- getEllipse(efit)
lines(e,col='red')

}
\references{
\item Radim Halir & Jan Flusser. 1998. Numerically stable direct least squares fitting of ellipses. Proceedings of the 6th International Conference in Central Europe on Computer Graphics and Visualization. WSCG '98, p. 125-132
}
\author{
JuG
}
