% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitJumpDistHist.r
\name{fitJumpDistHist}
\alias{fitJumpDistHist}
\title{Fit jump distribution histogram}
\usage{
fitJumpDistHist(data, nbPop = 1, initHist = list(D2 = 200, D1 = 0.1),
  timeRes = 1, breaks = 100, xlim, ylim, xlab, main, ...)
}
\arguments{
\item{data}{dataframe with jump}

\item{nbPop}{number of populations of diffusing species (1 or 2)}

\item{initHist}{initialisation parameters for the nls fit for exmaple list(D2 = 200,  D1=0.1) or list(D2 = .01,  D1=.1, D3=10, D4=100)}

\item{timeRes}{time resolution per unit of jump}

\item{breaks}{binning of the histogram (number of breaks)}
}
\description{
Fit jump distribution histogram
}
\examples{
xmlPath <- "/Users/jgodet/Seafile/MaBibliotheque/Code/TrackMate/nmeth.2808-sv1.xml"
data <- readTrackMateXML(XMLpath = xmlPath)
data$jump<-jump(data, spaceRes=1)
fitJumpDistHist(data=data, nbPop = 1, initHist = list(D2 = 200,  D1=0.1))
fitJumpDistHist(data=data, xlim=c(0,10), col=rgb(.2,.2,.2,.2), main="test")
# Results are affected by the histogram binning
fitJumpDistHist(data=data, breaks = 30)
fitJumpDistHist(data=data, breaks = 150)
fitJumpDistHist(data=data, nbPop=3)
fitJumpDistHist(data=data, nbPop=2, initHist = list(D2 = .01,  D1=.1, D3=10, D4=100))
}
\author{
JuG
}
