% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{frame_time}
\alias{frame_time}
\alias{frames_to_times}
\alias{times_to_frames}
\title{Convert between frame numbers and human readable time formats.}
\usage{
frames_to_times(frames, frame_rate)

times_to_frames(seconds, frame_rate)
}
\arguments{
\item{frames}{A numerc vector of frame numbers to convert.}

\item{frame_rate}{The frame rate of the tracking data in frames per second
(e.g. \code{tracks$params$frame_rate}).}

\item{seconds}{A character vector of times to convert (see examples).}
}
\description{
Convert between frame numbers and human readable time formats.
}
\section{Functions}{
\itemize{
\item \code{frames_to_times}: Convert frames to human-readable times.

\item \code{times_to_frames}: Convert human-readable times to frames.
}}
\examples{
# Using frame rate of 1 for clarity of usable formats
frames_to_times(1, 1)
frames_to_times(120, 1)
frames_to_times(3661, 1)

times_to_frames('5S', 1)
times_to_frames('2M 0S', 1)
times_to_frames('1H 1M 1S', 1)
times_to_frames('1H1M1S', 1)

# Note that the letters are not functional, just spaces work too:
times_to_frames('1 1 1', 1)
# Which means you can leave out larger denominations, but not smaller ones
times_to_frames('10S', 1) == times_to_frames('0M 10S', 1)
times_to_frames('2M', 1) == times_to_frames('2M 0S', 1)
}

