% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_track.R
\name{full_join_track}
\alias{full_join_track}
\title{full_join that shows how many rows match}
\usage{
full_join_track(x, y, by = NULL, suffix = c(".x", ".y"), .merge = FALSE, ...)
}
\arguments{
\item{x}{dataset}

\item{y}{dataset}

\item{by}{A character vector of variables to join by}

\item{suffix}{If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.}

\item{.merge}{If TRUE, then joined dataset will contain a variable called .merge that indicates which dataset the row came from}

\item{...}{}
}
\value{

}
\description{
full_join that shows how many rows match
}
