% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_paths.R
\name{future_path_pctls}
\alias{future_path_pctls}
\title{Generate Future Percentile Paths}
\usage{
future_path_pctls(
  data_fut,
  sequence_pctl = seq(20, 80, 20),
  changes_pctl,
  target_year = 2030,
  granularity = 0.1,
  min = NULL,
  max = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data_fut}{A data frame containing historical data with columns `code`, `year`, and `y`. This serves as the baseline for projections.}

\item{sequence_pctl}{Numeric vector of percentile paths to calculate.
The percentile score is determined by the granularity of the percentiles chosen here. For example, if seq(20,80,20) is chosen, then a country’s progress will fall in the 0th-20th percentile, 20th-40th percentile etc.
This argument is only relevant if `percentiles=TRUE`.}

\item{changes_pctl}{A `data.table` with predicted changes by initialvalue and pctl.}

\item{target_year}{An integer specifying the year to project to. Default is '2030'. Only relevant if `future = TRUE`.}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.
If `NULL`, predictions are unrestricted on the upper end.}

\item{min}{Optional. Minimum value of `initialvalue` to predict.}

\item{max}{Optional. Maximum value of `initialvalue` to predict.}

\item{verbose}{Logical; if `TRUE`, messages will be printed for each processing year. Default is `TRUE`.}
}
\value{
A data.table with the projected values (`y_fut`) by `code`, `year`, and `pctl`, including observed values where available.
}
\description{
Computes projected future paths for given percentiles based on initial levels and predicted changes.
}
\examples{
\dontrun{
data_fut <- data.frame(code = c("A", "A"), year = c(2020, 2021), y = c(50, 52))
predictions_pctl <- data.frame(initialvalue = 52, pctl = 20, change = 1)
future_path_pctls(data_fut = data_fut, predictions_pctl = predictions_pctl, target_year = 2025)
}

}
