% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWrapper.R
\name{gen_hash_id}
\alias{gen_hash_id}
\title{Create an 16 character "SpookyHashV2" hash of an R object.}
\usage{
gen_hash_id(x)
}
\arguments{
\item{x}{An R object}
}
\value{
An character vector with a prefix and 16 character hash of \code{x}
}
