% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R,
%   R/accessor-methods.R
\name{geomObject<-}
\alias{geomObject<-}
\alias{geomObject}
\alias{geomObject,PlotFeatureSet-method}
\alias{geomObject,ggplot-method}
\alias{geomObject,trellis-method}
\alias{geomObject,gTree-method}
\alias{geomObject<-,GGplotFeatureSet-method}
\alias{geomObject,GGplotFeatureSet-method}
\title{Accessors for the geom of GGplotFeatureSet objects}
\usage{
geomObject(object) <- value

geomObject(object)

\S4method{geomObject}{PlotFeatureSet}(object)

\S4method{geomObject}{ggplot}(object)

\S4method{geomObject}{trellis}(object)

\S4method{geomObject}{gTree}(object)

\S4method{geomObject}{GGplotFeatureSet}(object) <- value

\S4method{geomObject}{GGplotFeatureSet}(object)
}
\arguments{
\item{object}{An object of class GGplotFeatureSet. For accession, also plot objects of class ggplot.}

\item{value}{A named list of parameters for geometric objects in each layer of the plot.}
}
\value{
A modified object of class GGplotFeatureSet.
}
\description{
Get or set the geom associated with a GGplotFeatureSet object. This should not be called directly.
}
\examples{
library(ggplot2)
mt = datasets::mtcars
plt = qplot(mt$mpg, mt$cyl)
fs = makeFeatureSet(plt)
geomObject(fs)
}
