% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{get_scores}
\alias{get_scores}
\title{Wrapper to calculate percentile-based and/or speed-based scores}
\usage{
get_scores(
  speed = FALSE,
  pctl = FALSE,
  path_his_pctl = NULL,
  best = "high",
  path_his_speed = NULL,
  path_speed = NULL,
  min = 0,
  max = 100,
  granularity = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{speed}{Logical. If TRUE, compute speed-based score.}

\item{pctl}{Logical. If TRUE, compute percentile-based score.}

\item{path_his_pctl}{Data for `get_scores_pctl()`.}

\item{best}{"high" or "low", for `get_scores_pctl()`.}

\item{path_his_speed}{Data for `get_scores_speed()`.}

\item{path_speed}{Lookup for `get_scores_speed()`.}

\item{min}{Minimum value for both methods (default 0).}

\item{max}{Maximum value for both methods (default 100).}

\item{granularity}{Granularity for speed-based method (default 0.1).}
}
\value{
A list with `pctl` and/or `speed` elements (data.tables).
}
\description{
Calls `get_scores_pctl()` and/or `get_scores_speed()` depending on input.
}
