% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{get_scores_speed}
\alias{get_scores_speed}
\title{Calculate speed-based score over time}
\usage{
get_scores_speed(
  path_his_speed,
  path_speed,
  min = 0,
  max = 100,
  granularity = 0.1
)
}
\arguments{
\item{path_his_speed}{Result of `project_pctls_path()`}

\item{path_speed}{Result of `get_speed_path()`}

\item{min}{Minimum valid value for `y`. Default is 0.}

\item{max}{Maximum valid value for `y`. Default is 100.}

\item{granularity}{Rounding precision for `y`. Default is 0.1.}
}
\value{
A data.table with `code`, `score`, and `evaluationperiod`.
}
\description{
Computes a score based on the average annual change in percentile rank over at least 5 years.
}
