% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_changes.R
\name{get_speed_path}
\alias{get_speed_path}
\title{Get speed paths}
\usage{
get_speed_path(
  predictions_speed,
  granularity = 0.1,
  best = "high",
  verbose = TRUE
)
}
\arguments{
\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.}

\item{best}{Character string, either \code{"high"} (default) or \code{"low"}, indicating
whether higher or lower values of the initial variable are considered better.}
}
\value{
data frame with speed path
}
\description{
Calculates the expected path over time from the worst value to the best value.
}
