% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContour.r
\name{getContour}
\alias{getContour}
\title{Return the convexhull contour of coordinates}
\usage{
getContour(data, drawContour = FALSE, ...)
}
\arguments{
\item{data}{coordinate data}

\item{drawContour}{if TRUE draw a convexhull contour line}

\item{nbPix}{box size for the owin in ppp object}
}
\description{
Return the convexhull contour of corrdinates
}
\examples{


}
\author{
JuG
}
