% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEllipse.r
\name{getEllipse}
\alias{getEllipse}
\title{Create coordinates of an ellipse from fit ellipse parameters}
\usage{
getEllipse(fit, n = 360)
}
\arguments{
\item{fit}{a fit object from fitEllipse}

\item{n}{number of corrdinates points to build}
}
\value{

}
\description{
Create coordinates of an ellipse from fit ellipse parameters
}
\examples{


}
\author{
JuG
}
