% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{GGplotFeatureSet-class}
\alias{.GGplotFeatureSet}
\alias{GGplotFeatureSet-class}
\title{An S4 subclass of PlotFeatureSet with additional slots for objects of class ggplot.}
\section{Slots}{

\describe{
\item{\code{geom}}{A named list of parameters for geometric objects in each layer of the plot.}

\item{\code{stat}}{A named list of parameters for statistical transforms in each layer of the plot.}

\item{\code{position}}{A named list of positioning information in each layer of the plot.}

\item{\code{num.layers}}{An integer representing the number of layers in the plot.}
}}

