% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histJd.r
\name{histJd}
\alias{histJd}
\title{Plot JD histogram}
\usage{
histJd(jd, breaks, xlim, freq, pixSize, xlab, panel.first, ...)
}
\arguments{
\item{jd}{jump distribution vector}

\item{breaks}{histogram breaks}

\item{xlim}{xlim}

\item{pixSize}{pixel size in µm (if picels are not calibrated in trackMate)}

\item{xlab}{xlab}

\item{panel.first}{panel first options (use for exmaple grid())}
}
\value{
graph
}
\description{
Plot JD histogram (inherit from hist)
}
\examples{
histJd(jd, main="", panel.first=grid())
box()
}
\author{
JuG
}
