% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{vt_options}
\alias{vt_img_dir}
\alias{vt_img_ext}
\alias{vt_options}
\title{options extraction}
\usage{
vt_options(db)

vt_img_dir(db)

vt_img_ext(db)
}
\arguments{
\item{db}{Object to extract information from}
}

