% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_soc.R
\name{join_tr_to_soc}
\alias{join_tr_to_soc}
\title{Add select variables from the tr table to the soc table.}
\usage{
join_tr_to_soc(tracks, ...)
}
\arguments{
\item{tracks}{A tracks object.}

\item{...}{Which variables to join.}
}
\value{
A tracks object.
}
\description{
Use this function if you need access to track variables while computing
social variables. Each variable will be split into two columns, for fish1 and
fish2, with the numbers 1 and 2 amended. E.g. joining X and Y will generate
four new columns in \code{soc}, X1, X2, Y1, Y2.
}

