% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_track.R
\name{join_track}
\alias{join_track}
\title{Join two datasets and see how many matches there are}
\usage{
join_track(
  x,
  y,
  by = NULL,
  suffix = c(".x", ".y"),
  .merge = FALSE,
  join_type,
  ...
)
}
\arguments{
\item{x}{dataset}

\item{y}{dataset}

\item{by}{A character vector of variables to join by}

\item{suffix}{If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2.}

\item{.merge}{If TRUE, then joined dataset will contain a variable called .merge that indicates which dataset the row came from}

\item{join_type}{Which type of join? (full_join, left_join, right_join, inner_join)}

\item{...}{Other functions passed to the _join function}
}
\description{
Join two datasets and see how many matches there are
}
