% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{JSONBackend-class}
\alias{JSONBackend-class}
\alias{jsonbackend}
\title{JSON backend for trackr}
\section{Slots}{

\describe{
\item{\code{data}}{list. An in-memory list representation of the data in the db}

\item{\code{file}}{character. The file containing the db ( to read from and write to)}

\item{\code{last_load}}{POSIXct. The last time \code{data} was updated from disk.}
}}
\note{
This is a reference class, which does NOT have standard copy-on-write
semantics
}

