% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{JSONBackend}
\alias{JSONBackend}
\title{JSON backend contstructor}
\usage{
JSONBackend(file = normalizePath("./trackr_db_data.json"), data = list())
}
\arguments{
\item{file}{character. The json "database" to use as a trackr backend}
}
\value{
A JSONBackend object, for use in creating a TrackrDB object.
}
\note{
This function should generally not be called directly by end-users.
See instead \code{\link{jsonTDB}}
}
