% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{kbox}
\alias{kbox}
\title{Customized Cross-Entropy Clustering}
\usage{
kbox(
  x,
  centers = 1,
  iter.max = 10,
  split = FALSE,
  split.width = Inf,
  split.height = Inf,
  split.density = 0,
  min.size = 0,
  split.sensitivity = 0
)
}
\arguments{
\item{x}{A numeric matrix with two columns.}

\item{centers}{Either a matrix of initial centers or the number of initial
centers.}

\item{iter.max}{Maximum number of iterations at each clustering.}

\item{split}{Enables split mode. This mode discovers new clusters after
initial clustering, by trying to split single clusters into two.}

\item{split.width}{The maximum authorized width of a cluster. If a cluster is
wider than \code{split.width}, the function will attempt to split it in two.}

\item{split.height}{The maximum authorized height of a cluster. If a cluster
is higher than \code{split.height}, the function will attempt to split it in
two.}

\item{split.density}{The minimum authorized density of a cluster. If a
cluster is less dense than \code{split.density}, the function will attempt
to split it in two.}

\item{min.size}{The minimum authorized size (in number of items) of a cluster.
If a cluster is smaller than \code{min.size}, the function will attempt to
split it in two.}

\item{split.sensitivity}{The minimum amount of improvement in the cost
function of the cross-entropy clustering for a splitting event to be
considered valid.}
}
\value{
A matrix with 6 columns: x and y coordinates of the centers of the
 clusters, width, height, and angle of the covariance ellipse best describing
 each cluster, and the number of element in each cluster.
}
\description{
This function performs cross-entropy clustering on a data matrix.
 It is based on \code{\link[CEC]{cec}} but is limited to 2D matrices and
 implements its own splitting process.
}
\examples{
x <- c(rnorm(25, 4), rnorm(25, -2))
y <- c(rnorm(25, 2), rnorm(25, -3))
k <- kbox(cbind(x, y), 2)
plot(x, y, asp = 1)
apply(k, 1, function(k) {
  lines(ellipse(k[1], k[2], k[3], k[4], k[5]))
})

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
