% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{knit_and_record}
\alias{knit_and_record}
\title{Knit and record an Rmd, Rnw, etc file}
\usage{
knit_and_record(input, ..., verbose = FALSE, tmptdb = TrackrDB(backend =
  ListBackend()))
}
\arguments{
\item{input}{The input argument exactly as knitr's \code{knit} function
accepts it}

\item{...}{Passed directly to \code{knit}}

\item{verbose}{passed to (multiple) \code{record} calls for report and its
outputs}
}
\description{
This function wraps knitr's \code{knit} function in
a way that captures and records all the outputs printed in the final report,
associates them with the report, and then records the report as a whole.

This means that many records will generally be added to the trackr db for
a single call to this function.
}
\note{
as with all knitr support in the histry and trackr packages, manually
tracing certain functions within the knitr and evaluate packages will break
this function.
}
