% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{last_non_na}
\alias{last_non_na}
\title{Last non-NA in a vector
Calculate the last value in a vector that's not an NA, otherwise return NA}
\usage{
last_non_na(x)
}
\arguments{
\item{x}{vector}
}
\description{
Last non-NA in a vector
Calculate the last value in a vector that's not an NA, otherwise return NA
}
