% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_object.R
\name{load_tracks}
\alias{load_tracks}
\title{Load tracks object.}
\usage{
load_tracks(file)
}
\arguments{
\item{file}{File name.}
}
\value{
A tracks object.
}
\description{
This function allows for the reading of a tracks object from disk. The base
version of \code{load } cannot be used, since the tr and pairs tables need to
be reloaded onto a cluster.
}
\details{
In contrast to regular \code{load} in \code{base}, you should assign the
result.
}
\seealso{
save_tracks
}

