% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViztrackrDB-main-API.R
\name{record}
\alias{record}
\alias{rmEntry}
\alias{vtSearch}
\title{Primary high-level API functions for Viztrackr Databases}
\usage{
record(object, db = defaultVTDB(), code = histropts$history,
  force = FALSE, verbose = FALSE, symorpos = NULL)

rmEntry(object, db, verbose = FALSE)

vtSearch(pattern, fields = NULL, db, ret_type = c("doclist", "id",
  "backend"), verbose = FALSE)
}
\arguments{
\item{object}{ANY. A PlotFeatureSet (or plot object coercible to one) to be
added. Or (for rmPlot) the unique ID of an object in the database to remove.}

\item{db}{ViztrackrDB. The database}

\item{code}{ANY. Code/evaluation history to be associated with \code{object}}

\item{force}{logical. Overwrite any existing entry matching \code{object}.
(default: FALSE)}

\item{verbose}{logical. Should extra informative messages be displayed (
default: FALSE)}

\item{pattern}{character. A regular expression to match against the text in \code{fields}}

\item{fields}{character or NULL. The fiends in which to match, or NULL to
include all fields.}

\item{ret_type}{character. Format in which to return the response. Options are:
"id" - id of matching documents (default), "list" - A list containing the
matching documents represnted as R lists, and "backend" - a backend specific
representation of the set of matching documents}
}
\details{
These functions allow end-users to interact with
viztrackr databases. Each function does what its name suggests.
}

