% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-impls.R
\name{make_image_files}
\alias{make_image_files}
\title{Make image files for a featureset}
\usage{
make_image_files(object, opts)
}
\arguments{
\item{object}{A FeatureSet object}

\item{opts}{Options}
}
\value{
A named list with two entries: preview.path, and
    image.path. These should be paths to (now) existing iamge
    fiels for thumbnail and main display, respectively
}
\description{
Make image files for a featureset
}
