% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercerender.R
\name{manifestFromRecord}
\alias{manifestFromRecord}
\title{Generate a (switchr) Seed Manifest from a Record}
\usage{
manifestFromRecord(lst)
}
\arguments{
\item{lst}{The record in the form of a list (eg an element of the
list returned by findRecords}
}
\value{
A switchr SessionManifest object
}
\description{
Extracts the session info information within the
    record and uses it to generate a switchr manifest which can be
    used to reinstall exact versions of the R packages associated
    with the result
}
