% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{median_na}
\alias{median_na}
\title{Calculate median of a vector or return NA if there are no non-missing values to calculate from
Useful for summarising}
\usage{
median_na(x)
}
\arguments{
\item{x}{vector}
}
\description{
Calculate median of a vector or return NA if there are no non-missing values to calculate from
Useful for summarising
}
