% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{min_na}
\alias{min_na}
\title{Calculate max of a vector or return NA if there are no non-missing values to calculate from
Useful for summarising}
\usage{
min_na(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Calculate max of a vector or return NA if there are no non-missing values to calculate from
Useful for summarising
}
