% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{moduleSVR}
\alias{moduleSVR}
\title{Source Module Server}
\usage{
moduleSVR(module)
}
\arguments{
\item{module}{The module name in the form "app/module".}
}
\value{
The server code of the module.
}
\description{
Utility function to load the different server modules of the apps.
}
\examples{
\dontrun{
moduleSVR("trackR/video")
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
