% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{moduleUI}
\alias{moduleUI}
\title{Source Module UI}
\usage{
moduleUI(module)
}
\arguments{
\item{module}{The module name in the form "app/module".}
}
\value{
The UI code of the module.
}
\description{
Utility function to load the different UI modules of the apps.
}
\examples{
\dontrun{
moduleUI("trackR/video")
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
