% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_soc.R
\name{mutate_soc}
\alias{mutate_soc}
\alias{nip_dist}
\alias{nip_dist_}
\alias{pair_dist}
\alias{pair_dist_}
\title{Conveniently add common new variables to the \code{soc} table.}
\usage{
pair_dist(x1 = X1, x2 = X2, y1 = Y1, y2 = Y2)

pair_dist_(x1, x2, y1, y2)

nip_dist(x1 = X1, x2 = X2, y1 = Y1, y2 = Y2,
  minor_axis1 = minor_axis1, minor_axis2 = minor_axis2,
  major_axis1 = major_axis1, major_axis2 = major_axis2,
  orientation1 = orientation1, orientation2 = orientation2, n = 20)

nip_dist_(x1, x2, y1, y2, minor_axis1, minor_axis2, major_axis1, major_axis2,
  orientation1, orientation2, n = 20)
}
\arguments{
\item{x1}{X-coordinate for animal 1.}

\item{x2}{X-coordinate for animal 2.}

\item{y1}{Y-coordinate for animal 1.}

\item{y2}{Y-coordinate for animal 2.}

\item{minor_axis1}{Minor axis (*b*) of the ellipse fit of animal 1.}

\item{minor_axis2}{Minor axis (*b*) of the ellipse fit of animal 2.}

\item{major_axis1}{Major axis (*a*) of the ellipse fit of animal 1.}

\item{major_axis2}{Major axis (*a*) of the ellipse fit of animal 2.}

\item{orientation1}{Angle of ellipse fit with the x-axis for animal 1.}

\item{orientation2}{Angle of ellipse fit with the x-axis for animal 2.}

\item{n}{Number of point used for numerical approximation.}
}
\description{
These functions are designed for use within \code{summarise_}, and will
compute common parameters for pairs of animals. By default, these functions
will use the expected variable names, as are default in \code{tracks}
objects, but they can be overridden.
}
\details{
Many of these functions rely on parameters that originate from the \code{tr}
table, such as \code{X1} or \code{orientation2}. This usually means you have
to add those variable to the \code{soc} table first, by calling
\code{join_tr_to_soc}.
}
\seealso{
mutate_tr mutate_.tracks
}

