% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_tr.R
\name{mutate_tr}
\alias{acceleration}
\alias{acceleration_}
\alias{change}
\alias{change_}
\alias{distance}
\alias{distance_}
\alias{heading}
\alias{heading_}
\alias{mutate_tr}
\alias{speed}
\alias{speed_}
\alias{turn}
\alias{turn_}
\title{Conveniently add common new variables to the \code{soc} table.}
\usage{
change(x, order_by = frame, n = 1)

change_(x, order_by, n = 1)

distance(x = X, y = Y, order_by = frame)

distance_(x, y, order_by)

speed(x = X, y = Y, order_by = frame)

speed_(x, y, order_by)

acceleration(x = X, y = Y, order_by = frame)

acceleration_(x, y, order_by)

heading(x = X, y = Y, order_by = frame)

heading_(x, y, order_by)

turn(x = X, y = Y, order_by = frame)

turn_(x, y, order_by)
}
\arguments{
\item{x}{X-coordinate.}

\item{order_by}{This parameter controls the ordering of \code{dplyr::lag} and
\code{dplyr::lead} and it is strongly advised to leave it at it's default
(\code{frame}).}

\item{y}{Y-coordinate.}
}
\description{
These functions are designed for use within \code{summarise_}, and will
compute common derived parameters for tracks. By default, these functions
will use the expected variable names, as are default in \code{tracks}
objects, but they can be overridden.
}
\seealso{
mutate_soc mutate_.tracks
}

