% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_track.R
\name{mutate_track}
\alias{mutate_track}
\title{Track what changes are made when you mutate}
\usage{
mutate_track(.data, ...)
}
\arguments{
\item{.data}{dataset}

\item{...}{arguments to pass to mutate}
}
\value{

}
\description{
Track what changes are made when you mutate
}
