% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{ndoc,TrackrDB-method}
\alias{ndoc,TrackrDB-method}
\alias{ndoc-method,TrackrDB}
\title{Number of docs in a TrackrDB
Query the backend for the number of docs it contains}
\usage{
\S4method{ndoc}{TrackrDB}(x, ...)
}
\arguments{
\item{x}{TrackrDB.}

\item{...}{not used.}
}
