% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{ndoc,ViztrackrDB-method}
\alias{ndoc,ViztrackrDB-method}
\alias{ndoc-method,ViztrackrDB}
\title{Number of docs in a vtdb
Query the backend for the number of docs it contains}
\usage{
\S4method{ndoc}{ViztrackrDB}(x, ...)
}
\arguments{
\item{x}{ViztrackrDB.}

\item{...}{not used.}
}

