% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_soc.R
\name{nip_dist}
\alias{nip_dist}
\title{Add pairwise distance based on tip to ellipse.}
\usage{
nip_dist(x1, x2, y1, y2, minor_axis1, minor_axis2, major_axis1, major_axis2,
  orientation1, orientation2, n = 20)
}
\arguments{
\item{n}{How many points along the ellipse are used for the approximation.}

\item{tracks}{A tracks object.}
}
\value{
A tracks object
}
\description{
Add the distance between the pair of animals for each entry in the $soc
table. This distance is from the mouth of fish 1 (tip of the ellipse) to
an approximately closest point on the ellipse for fish 2.
}
\details{
It does not get the closest point, as this calculation is quite involved, but
does a numerical approximation based on n points along the ellipse.
}

