% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_soc.R
\name{nip_dist_velocity}
\alias{nip_dist_velocity}
\title{Calculate change in pairwise distance based on tip to ellipse.}
\usage{
nip_dist_velocity(x1, x2, y1, y2, minor_axis1, minor_axis2, major_axis1,
  major_axis2, orientation1, orientation2, n = 20, f)
}
\arguments{
\item{n}{How many points along the ellipse are used for the approximation.}
}
\value{
A vector of velocities in px per frame.
}
\description{
Calculate change in pairwise distance based on tip to ellipse.
}

