% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R, R/accessor-methods.R
\docType{methods}
\name{nLayers<-}
\alias{nLayers}
\alias{nLayers,GGplotFeatureSet-method}
\alias{nLayers,PlotFeatureSet-method}
\alias{nLayers,gTree-method}
\alias{nLayers,ggplot-method}
\alias{nLayers,trellis-method}
\alias{nLayers<-}
\alias{nLayers<-,GGplotFeatureSet-method}
\title{Mutator method for GGplotFeatureSet "num.layers" slot.}
\usage{
nLayers(object) <- value

nLayers(object)

\S4method{nLayers}{PlotFeatureSet}(object)

\S4method{nLayers}{ggplot}(object)

\S4method{nLayers}{trellis}(object)

\S4method{nLayers}{gTree}(object)

\S4method{nLayers}{GGplotFeatureSet}(object) <- value

\S4method{nLayers}{GGplotFeatureSet}(object)
}
\arguments{
\item{object}{An object of class GGplotFeatureSet. For accession, also plot objects of class ggplot.}

\item{value}{An integer representing the number of layers in the plot.}
}
\value{
A modified object of class GGplotFeatureSet.

An integer representing the number of layers in the plot.
}

