% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWrapper.R
\name{not_implemented}
\alias{not_implemented}
\title{Internal utility function to warn the user that a requested feature is not yet implemented.}
\usage{
not_implemented(reason = NULL)
}
\arguments{
\item{reason}{An optional character vector describing the feature.}
}

