% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/accessor-methods.R
\docType{methods}
\name{objCode<-}
\alias{objCode}
\alias{objCode,ObjFeatureSet-method}
\alias{objCode<-}
\alias{objCode<-,ObjFeatureSet-method}
\title{Mutator method for "code" slot of PlotFeatureSet (or one of its subclasses).}
\usage{
objCode(object) <- value

objCode(object)

\S4method{objCode}{ObjFeatureSet}(object) <- value

\S4method{objCode}{ObjFeatureSet}(object)
}
\arguments{
\item{object}{An object of (super)class PlotFeatureSet. For accession, also plot objects of class ggplot or trellis.}

\item{value}{A CodeDepends::Script object.}
}
\value{
A modified object of (super)class PlotFeatureSet.

A CodeDepends::Script object
}

