% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWrapper.R
\name{ggplot-class}
\alias{ggplot-class}
\alias{gg}
\alias{gg-class}
\alias{trellis}
\alias{trellis-class}
\alias{gTree}
\alias{gTree-class}
\title{ggplot S4 classes}
\description{
formal S4 classes for S3 classes defined in other packages.
}
