% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_soc.R
\name{pair_dist}
\alias{pair_dist}
\title{Calculate pairwise distance.}
\usage{
pair_dist(x1, x2, y1, y2)
}
\arguments{
\item{x1}{x-coordinate for fish 1.}

\item{x2}{x-coordinate for fish 2.}

\item{y1}{y-coordinate for fish 1.}

\item{y2}{y-coordinate for fish 2.}
}
\value{
A vector of distances in pixels.
}
\description{
Calculate the distance between centroids for a pair of animals.
}

