% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_soc.R
\name{pair_dist_acceleration}
\alias{pair_dist_acceleration}
\title{Calculate accereleration in distance between pairs.}
\usage{
pair_dist_acceleration(x1, x2, y1, y2, f)
}
\arguments{
\item{x1}{x-coordinate for fish 1.}

\item{x2}{x-coordinate for fish 2.}

\item{y1}{y-coordinate for fish 1.}

\item{y2}{y-coordinate for fish 2.}

\item{f}{Always use 'f' or 'frame' (equivalent). Used to order by frame
before calculation.}
}
\value{
A vector of accelerations in pixels per frame^2.
}
\description{
Calculate the change in relative velocity between the pair of animals.
Negative values mean the animals are decreasing the rate at which their
distance changes, postitive values means an increase.
}

