% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_soc.R
\name{pair_dist_velocity}
\alias{pair_dist_velocity}
\title{Calculate change in distance between pairs.}
\usage{
pair_dist_velocity(x1, x2, y1, y2, f)
}
\arguments{
\item{x1}{x-coordinate for fish 1.}

\item{x2}{x-coordinate for fish 2.}

\item{y1}{y-coordinate for fish 1.}

\item{y2}{y-coordinate for fish 2.}

\item{f}{Always use 'f' or 'frame' (equivalent). Used to order by frame
before calculation.}
}
\value{
A vector of velocities in pixels per frame.
}
\description{
Calculate the change in distance between the pair of animals. Negative values
mean the animals are moving closer to each other, postitive values means they
are moving apart.
}

