% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/parseCode.R
\docType{methods}
\name{parseCode}
\alias{parseCode}
\alias{parseCode,expression-method}
\alias{parseCode,call-method}
\alias{parseCode,character-method}
\alias{parseCode,ggplot-method}
\alias{parseCode,trellis-method}
\alias{parseCode,ANY-method}
\title{Parse result creation code}
\usage{
parseCode(code)

\S4method{parseCode}{expression}(code)

\S4method{parseCode}{call}(code)

\S4method{parseCode}{character}(code)

\S4method{parseCode}{ggplot}(code)

\S4method{parseCode}{trellis}(code)

\S4method{parseCode}{ANY}(code)
}
\arguments{
\item{code}{An expression, call or character vector, or a PlotFeatureSet or plot object. If missing, this method will attempt to extract code out of the plot object.}
}
\value{
A CodeDepends::Script object.
}
\description{
parse code into a CodeDepends::Script
}
