% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjFeatureSet.R
\name{pfs_list_collapse}
\alias{pfs_list_collapse}
\title{Collapse particular fields of a ObjFeatureSet to delimited character vectors.}
\usage{
pfs_list_collapse(pfs)
}
\arguments{
\item{pfs}{An object of (super)class ObjFeatureSet.}
}
\value{
A list.
}
