% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWrapper.R
\docType{methods}
\name{plot,PlotFeatureSet,missing-method}
\alias{plot,PlotFeatureSet,missing-method}
\title{Display the plot object owned by PlotFeatureSet objects and subclasses thereof.}
\usage{
\S4method{plot}{PlotFeatureSet,missing}(x, y, ...)
}
\arguments{
\item{x}{An object of (super)class PlotFeatureSet.}

\item{y}{Should be left empty.}

\item{...}{Other named arguments (currently unused).}
}
\value{
The plot object, plotted through its native print method.
}
