% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_plots.R
\name{plot.tracks}
\alias{plot.tracks}
\title{Plot paths}
\usage{
\method{plot}{tracks}(x, color = ~animal, facet = ~trial, nrow = NULL,
  ncol = NULL, ...)
}
\arguments{
\item{x}{A tracks object}

\item{color}{Color specification of the path as a formula (e.g. ~animal).}

\item{facet}{Facet specification as a formula (e.g. ~trial), passed to
facet_wrap.}

\item{nrow}{Control number of rows for the facets.}

\item{ncol}{Control number of columns for the facets.}

\item{...}{Ignored.}
}
\value{
A ggplot object
}
\description{
Only plots a path between subsequent frames.
}
\examples{
plot(guppies)
plot(guppies, NULL) # No animal coloring
plot(guppies, facet = ~1) # No facets
plot(guppies, facet = ~trial + animal, ncol = 8) # more complex facetting
}

