% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_plots.R
\name{plot_lag_cor}
\alias{plot_lag_cor}
\title{Plot lag correlations}
\usage{
plot_lag_cor(data)
}
\arguments{
\item{data}{The result from a lag correlation function (e.g.
\code{calc_speed_lag}).}
}
\value{
A \code{ggplot} object.
}
\description{
This function produces a \code{ggplot} that shows how the lag correlations
changes over time for each trial. Tries to do clever facetting. Will show
all pairs in the data, but if all trials contain only one pair, will only
show pair 1,2 (and not 2,1).
}
\details{
The y-axis shows time bins (top to bottom), the x-axis shows the lag in
frames and the size (area) of the dots denotes strength of the correlation
(*r*).
}

