% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_base_methods.R, R/tracks_plots.R
\name{plot.tracks}
\alias{plot.tracks}
\alias{plot_tracks}
\title{Plot paths}
\usage{
\method{plot}{tracks}(x, ...)

plot_tracks(tracks, color = ~animal, facet = ~trial, nrow = NULL,
  ncol = NULL)
}
\arguments{
\item{x}{A tracks object.}

\item{...}{Pass arguments to \code{plot_tracks}.}

\item{tracks}{A tracks object.}

\item{color}{Color specification of the path as a formula (e.g. ~animal).}

\item{facet}{Facet specification as a formula (e.g. ~trial), passed to
facet_wrap.}

\item{nrow}{Control number of rows for the facets.}

\item{ncol}{Control number of columns for the facets.}
}
\value{
A \code{ggplot} object.
}
\description{
Only plots a path between subsequent frames.
}
\examples{
Guppies <- as_tracks(guppies, 30, 1080)
plot(Guppies)
# No animal coloring
plot(Guppies, color = NULL)
# Plot trials in the same panel
plot(Guppies, color = ~animal:trial,facet = ~1)
# more complex facetting
plot(Guppies, facet = ~trial + animal, ncol = 8)
}
