% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_plots.R
\name{plot_tracks_quick}
\alias{plot_tracks_quick}
\title{Quickly plot a tracks object.}
\usage{
plot_tracks_quick(tracks, ...)
}
\arguments{
\item{tracks}{A tracks object.}

\item{...}{Other arguments passed on to \code{plot}.}
}
\value{
Nothing.
}
\description{
No facetting available.
}
\examples{
Guppies <- as_tracks(guppies, 30, 1080)
plot_tracks_quick(Guppies)
}
