% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiffusionMap.r
\name{plotDiffusionMap}
\alias{plotDiffusionMap}
\title{Do something}
\usage{
plotDiffusionMap(azc2, nbPix = 2, jdRange = c(0, 12),
  colRange = NULL, ...)
}
\arguments{
\item{azc2}{dataframe with x, y, track, jd and cluster variable}

\item{nbPix}{boxsize (expressed in x or y units)}

\item{jdRange}{jump distance limit (ignore all jumps larger than jdlim)}

\item{colRange}{rage of jump distance}
}
\description{
Do something
}
\examples{

dataTest <- expData[expData$cluster \%in\% "PvdA9_8",]
plotDiffusionMap(dataTest, nbPix=.5 )
gcont <- getContour(dataTest[, c("x","y")],drawContour = FALSE)
drawRod(gcont,lty=3)
}
\author{
JuG
}
