% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{PlotFeatureSet-class}
\alias{PlotFeatureSet-class}
\alias{.PlotFeatureSet}
\title{An S4 class called PlotFeatureSet}
\section{Slots}{

\describe{
\item{\code{data}}{A list of data.frames containing the variables and observations used in the plot.}

\item{\code{titles}}{Title and subtitle of the plot object; a named list of the form list(main = "My title", sub = "My subtitle").}

\item{\code{varlabels}}{Variable labels of the plot object; a named list of the form list(x = "X axis label", y = "Y axis label", groups = list(...)). Note that non-empty labels are character vectors and may contain more than one entry.}

\item{\code{annotation.text}}{Annotation text of the plot object.}

\item{\code{vartypes}}{Variable types of the plot object; a named list of the form list(x = "numeric", y = "factor", ...).}

\item{\code{grouping}}{Grouping information on the plot object; a named list.}

\item{\code{coordsys}}{A character vector describing the coordinate system employed in the plot.}

\item{\code{nobs}}{An integer representing the number of observations in the plotted data.}

\item{\code{haslegend}}{A boolean indicating whether or not a legend is displayed in the plot.}

\item{\code{tags}}{A character vector of user-defined tags.}

\item{\code{code}}{R code to reproduce the plot, as a CodeDepends::Script object. May be empty.}

\item{\code{codeinfo}}{Information about the R code to reproduce the plot, as a CodeDepends::ScriptInfo object. May be empty.}
}}

