% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMSD.r
\name{plotMSD}
\alias{plotMSD}
\title{Do something}
\usage{
plotMSD(msdData, deltaT, fitMSD = TRUE, printMSDfit = TRUE, ...)
}
\arguments{
\item{msdData}{MSD data (list) (output of calcMSD function)}

\item{deltaT}{time elapsed between two consecutive frame}

\item{fitMSD}{boolean}

\item{printMSDfit}{boolean}
}
\description{
Do something
}
\examples{


}
\author{
JuG
}
