% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R, R/accessor-methods.R
\docType{methods}
\name{position<-}
\alias{position}
\alias{position,GGplotFeatureSet-method}
\alias{position,PlotFeatureSet-method}
\alias{position,gTree-method}
\alias{position,ggplot-method}
\alias{position,trellis-method}
\alias{position<-}
\alias{position<-,GGplotFeatureSet-method}
\title{Mutator method for GGplotFeatureSet "position" slot.}
\usage{
position(object) <- value

position(object)

\S4method{position}{PlotFeatureSet}(object)

\S4method{position}{ggplot}(object)

\S4method{position}{trellis}(object)

\S4method{position}{gTree}(object)

\S4method{position}{GGplotFeatureSet}(object) <- value

\S4method{position}{GGplotFeatureSet}(object)
}
\arguments{
\item{object}{An object of class GGplotFeatureSet. For accession, also plot objects of class ggplot.}

\item{value}{A named list of positioning information in each layer of the plot.}
}
\value{
A modified object of class GGplotFeatureSet.

A named list of positioning information in each layer of the plot.
}

