% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_changes.R
\name{predict_changes}
\alias{predict_changes}
\title{Predict Changes in an indicator Over Time}
\usage{
predict_changes(
  data,
  min = NULL,
  max = NULL,
  floor = 0,
  ceiling = 100,
  granularity = 0.1,
  lambdas = NULL,
  best = "high",
  speed = FALSE,
  percentiles = TRUE,
  pctlseq = seq(20, 80, 20),
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame or list representing the input data model to be used for the prediction}

\item{min}{Optional. Minimum value of `initialvalue` to predict. If `NULL`,
it is set to `floor` (if provided) or `min(data_model$initialvalue)`.}

\item{max}{Optional. Maximum value of `initialvalue` to predict. If `NULL`,
it is set to `ceiling` (if provided) or `max(data_model$initialvalue)`.}

\item{floor}{Numeric or `NULL`.Minimum value of indicator.
If `NULL`, predictions are unrestricted on the lower end.}

\item{ceiling}{Maximum value of indicator (NA if none).
If `NULL`, predictions are unrestricted on the upper end.}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.}

\item{lambdas}{Optional. A vector of lambda values, i.e., levels of flexibility that should be tried and evaluated through cross validation
Defaults to `0.1 * 1.148^(0:50)`.}

\item{best}{Character string, either \code{"high"} (default) or \code{"low"}, indicating
whether higher or lower values of the initial variable are considered better.}

\item{speed}{Logical. If TRUE, calculate speed of progress scores.
Default is FALSE}

\item{percentiles}{Logical. If TRUE, calculate percentile scores. Default is TRUE}

\item{pctlseq}{Percentiles to calculate. The percentile score assigned is determined by the granularity of the percentiles selected here. I.e., if seq(20,80,20) is selected, then we will be able to tell if a country is between 0-20, 20-40, etc.}
}
\value{
A list containing prediction results. The content depends on the selected method:
\describe{
  \item{`predictions_speed`}{Predicted changes using the `"speed"` method (if selected).}
  \item{`path_speed`}{The path or trajectory of predictions based on the `"speed"` method (if selected).}
  \item{`predictions_pctls`}{Predicted changes using the `"percentiles"` method (if selected).}
}
}
\description{
This function calculates projected changes in an indicator based on its initial value.
Two methods are available: `"speed"` and `"percentiles"`.
}
\details{
- On The `"speed"` method: for indicators without overall progress, the speed of progress cannot be computed.
- The `"percentiles"` method can be computed for all indicators but ideally require rich training data.
}
\seealso{
[predict_speed()], [get_speed_path()], [predict_pctls()]
}
