% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_changes.R
\name{predict_pctls}
\alias{predict_pctls}
\title{Predict percentiles}
\usage{
predict_pctls(
  data_model,
  min = NULL,
  max = NULL,
  granularity = 0.1,
  pctlseq = seq(20, 80, 20),
  floor = 0,
  ceiling = 100,
  verbose = TRUE
)
}
\arguments{
\item{data_model}{A data frame with the training data. Must contain:
- `initialvalue`: numeric, the initial level of the indicator.
- `change`: numeric, the change in the indicator.
- `fold_id`: numeric or factor, used for cross-validation.}

\item{min}{Optional. Minimum value of `initialvalue` to predict. If `NULL`,
it is set to `floor` (if provided) or `min(data_model$initialvalue)`.}

\item{max}{Optional. Maximum value of `initialvalue` to predict. If `NULL`,
it is set to `ceiling` (if provided) or `max(data_model$initialvalue)`.}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.}

\item{pctlseq}{Percentiles to calculate. The percentile score assigned is determined by the granularity of the percentiles selected here. I.e., if seq(20,80,20) is selected, then we will be able to tell if a country is between 0-20, 20-40, etc.}

\item{floor}{Numeric or `NULL`.Minimum value of indicator.
If `NULL`, predictions are unrestricted on the lower end.}

\item{ceiling}{Maximum value of indicator (NA if none).
If `NULL`, predictions are unrestricted on the upper end.}
}
\value{
A data frame with calculated changes by percentiles.
}
\description{
This function calculates the changes as a function of initial values by percentile
}
