% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWrapper.R
\name{print.gTree}
\alias{print.gTree}
\title{Print method for an object of class gTree.}
\usage{
\method{print}{gTree}(x, ...)
}
\arguments{
\item{x}{An object of class gTree.}

\item{...}{Other named arguments (currently unused).}
}
\description{
print a gTree object.
}
