% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_paths.R
\name{project_path_speed}
\alias{project_path_speed}
\title{Historical speed path}
\usage{
project_path_speed(
  data_his,
  sequence_speed = c(0.25, 0.5, 1, 2, 4),
  path_speed,
  granularity = 0.1,
  eval_from = 2000,
  eval_to = 2022,
  min = NULL,
  max = NULL,
  best = "high"
)
}
\arguments{
\item{data_his}{A `data.table` containing historical values with variables `code`, `year`, and `y_his`.}

\item{sequence_speed}{numeric vector of speed paths to calculate}

\item{path_speed}{Result of `get_speed_path()` — the base simulated trajectories}

\item{granularity}{Rounding precision for `y_speed`. Default is 0.1.}

\item{eval_from}{First year to include when evaluating the progress of countries.}

\item{eval_to}{Last year to include when evaluating the progress of countries.}

\item{min}{Numeric (optional). Minimum value of the indicator. If NULL, the minimum is computed from the data and rounded to the nearest multiple of `granularity`.
Will be overruled by support if that reflects a higher minimum value}

\item{max}{Numeric (optional). Maximum value of the indicator. If NULL, the maximum is computed from the data and rounded to the nearest multiple of `granularity`. Will be overruled by support if that reflects a lower maximum value.}

\item{best}{A character string indicating whether higher or lower values of thei indicator are better. Use "high" if higher values of the indicator reflect better outcomes, or "low" if lower values are better.}

\item{path_his_speed}{Result of `project_path_speed()` — contains simulated paths (`y_speed`)}
}
\value{
A data table of projected values under different speeds



Calculate speed-based score over a historical evaluation period

Computes countries’ speed of progress over a user-specified evaluation period.

A data.table with `code`, `score`, and `evaluationperiod`.
}
\description{
Calculates path a country would have taken with various speeds
}
