% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_paths.R
\name{project_pctls_path}
\alias{project_pctls_path}
\title{Project percentiles paths}
\usage{
project_pctls_path(
  data_his,
  start_year = 2000,
  end_year = 2022,
  granularity = 0.1,
  floor = 0,
  ceiling = 100,
  min = NULL,
  max = NULL,
  pctlseq = seq(20, 80, 20),
  predictions_pctl,
  verbose = TRUE
)
}
\arguments{
\item{data_his}{A `data.table` containing historical values with variables `code`, `year`, and `y_his`.}

\item{start_year}{The first year to include in the analysis}

\item{end_year}{The last year to include in the analysis}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.}

\item{floor}{Numeric or `NULL`.Minimum value of indicator.
If `NULL`, predictions are unrestricted on the lower end.}

\item{ceiling}{Maximum value of indicator (NA if none).
If `NULL`, predictions are unrestricted on the upper end.}

\item{min}{Optional. Minimum value of `initialvalue` to predict. If `NULL`,
it is set to `floor` (if provided) or `min(data_model$initialvalue)`.}

\item{max}{Optional. Maximum value of `initialvalue` to predict. If `NULL`,
it is set to `ceiling` (if provided) or `max(data_model$initialvalue)`.}

\item{pctlseq}{Numeric vector. Sequence of percentiles}

\item{predictions_pctl}{A `data.table` with predicted changes by `initialvalue` and `pctl`.}

\item{verbose}{Logical. Whether to print progress messages.}
}
\value{
A `data.table` with projected values `y_his` by `code`, `year`, and `pctl`.
}
\description{
Simulates year-by-year projected paths of a variable across percentiles,
based on historical values and predicted changes.
}
