% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_idTracker}
\alias{read_idTracker}
\title{Read idTracker data}
\usage{
read_idTracker(file = NULL, folder = NULL, type = "nogaps",
  multiple = TRUE, animals = NULL)
}
\arguments{
\item{file}{Either directly supply a path to a file, a list of paths,}

\item{folder}{Or supply a path to a folder.}

\item{type}{If folder is supplied, use type to indicate which track type
should be read. Currently always tries to find nogaps files.}

\item{multiple}{If TRUE, will try to find files in the subfolders of folder
and combine them into one long data.frame. Only use when using folder.}

\item{animals}{Optional vector of names (ordered in the same order as the
tracking id's). Otherwise numbers are used instead.}
}
\value{
A long \code{data.frame} (with class \code{tbl_df} and source
\code{idTracker}).
}
\description{
Loads a txt file containing tracking data from idTracker into R.
}
\details{
This function generally assumes that each file is in it's own folder. If not,
build a list of file paths beforehand and supply that as file.
}
\seealso{
Other trackr.read.functions: \code{\link{read_Ctrax}}
}

