% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrackMateXML.r
\name{readTrackMateXML}
\alias{readTrackMateXML}
\title{read TrackMate XML output files}
\usage{
readTrackMateXML(XMLpath)
}
\arguments{
\item{XMLpath}{path to the xml file}
}
\description{
read TrackMate XML output files
}
\examples{
xmlPath <- "/Users/jgodet/Desktop/sv1.xml"
readTrackMateXML(XMLpath = xmlPath)

}
\author{
JuG
}
