% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrackMateXmlTrace.r
\name{readTrackMateXmlTrace}
\alias{readTrackMateXmlTrace}
\title{read TrackMate XML traces output files}
\usage{
readTrackMateXmlTrace(XMLpath, timeRes = 1)
}
\arguments{
\item{XMLpath}{XML file path}

\item{timeRes}{time exposure per frame}
}
\value{

}
\description{
Function to read TrackMate XML traces output files: much faster than reading the complete TRack Mate XML file
}
\examples{
xmlPath <- "/Users/jgodet/Seafile/MaBibliotheque/Code/TrackMate/nmeth.2808-sv1Trace.xml"
data <- readTrackMateXmlTrace(XMLpath = xmlPath)

data$jump<-jump(data, spaceRes=1)
hist(data$jump, breaks=150)
fitJumpDistHist(data=data, breaks = 150)
msdMat <- calcMSD(data, N=2)
plotMSD(msdMat, col='blue')

}
\author{
JuG
}
