% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R
\docType{methods}
\name{regDateTime<-}
\alias{regDateTime}
\alias{regDateTime,ObjFeatureSet-method}
\alias{regDateTime<-}
\alias{regDateTime<-,ObjFeatureSet-method}
\title{Mutator method for "regdate" slot of ObjFeatureSet (or one of its subclasses).}
\usage{
regDateTime(object) <- value

regDateTime(object)

\S4method{regDateTime}{ObjFeatureSet}(object) <- value

\S4method{regDateTime}{ObjFeatureSet}(object)
}
\arguments{
\item{object}{An object of (super)class ObjFeatureSet.}

\item{value}{A POSIXct timestamp representing the moment when the plot was registered.}
}
\value{
A modified object of (super)class ObjFeatureSet.

A POSIXct timestamp representing the moment when the plot was registered.
}

