% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{replace_with_na}
\alias{replace_with_na}
\title{Replace with NA if certain value}
\usage{
replace_with_na(x, values)
}
\arguments{
\item{x}{vector}

\item{values}{values to replace with NA}
}
\value{

}
\description{
Replace with NA if certain value
}
