% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{roll}
\alias{roll}
\title{Roll a function over a tracks variable.}
\usage{
roll(tracks, table = "tr", var, fun = mean, window = 5, weights = NULL)
}
\arguments{
\item{tracks}{A tracks object.}

\item{table}{Which table to look up the variable.}

\item{var}{Which variable to roll over.}

\item{fun}{Which function to use.}

\item{window}{Set to \code{NULL} if it has to be ignored. Ignored if weights
is non-\code{NULL}.}

\item{weights}{Alternatively to window, you can supply a set of weights.
A good example: \code{dnorm(seq(-3, 3, length.out = 20))}.}
}
\value{
A tracks object.
}
\description{
This is mainly useful for smoothing a variable over time.
}
\details{
For the common operations of mean, median, min, max, sum, prod, sd and var,
fast implementations from the \code{RcppRoll} package are used.
}

