% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{RStudioExtras}
\alias{RStudioExtras}
\alias{RStudioExtras-class}
\title{RStudioExtras class}
\description{
A class to contain information about the current session when
working inside the RStudio IDE.
}
