% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_object.R
\name{save_tracks}
\alias{save_tracks}
\title{Save tracks object.}
\usage{
save_tracks(tracks, file)
}
\arguments{
\item{tracks}{A tracks object}

\item{file}{File name.}
}
\value{
invisible(NULL)
}
\description{
This function allows for the writing of a tracks object to disk. The base
version of \code{save } cannot be used, since the tr and pairs tables need to
be pulled of the cluster.
}
\details{
Name will
}
\seealso{
load_tracks
}

