% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/ObjFeatureSet.R
\docType{methods}
\name{selectPlot}
\alias{selectPlot}
\alias{selectPlot,ObjFeatureSet-method}
\title{Search the features of a PlotFeatureSet.}
\usage{
selectPlot(object, terms, ...)

\S4method{selectPlot}{ObjFeatureSet}(object, terms, features = NULL,
  improvise = TRUE, verbose = TRUE)
}
\arguments{
\item{object}{A PlotFeatureSet.}

\item{terms}{The terms for which the algorithm should search.}

\item{...}{Other named arguments.}

\item{features}{A character vector naming the features which should be searched. All by default.}

\item{improvise}{A boolean indicating whether the algorithm should automatically expand the search to all features if the search is fruitless in the specified features. True by default.}

\item{verbose}{A boolean indicating whether match information should be printed. True by default.}
}
\value{
A boolean indicating whether the terms were found.
}
