% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_trial_operation.R
\name{set_start_frame}
\alias{set_start_frame}
\title{Set a new starting frame}
\usage{
set_start_frame(tracks, value = NULL, tab = NULL, frame = "frame",
  trial = "trial")
}
\arguments{
\item{tracks}{A tracks object.}

\item{value}{A single value to set as the start frame.}

\item{tab}{A data.frame with trial names and start frames. Use \code{frame}
and \code{trial} to choose which variables are used.}

\item{frame}{The column from \code{tab} that is used as the starting frame.
By default this is \code{'frame'}.}

\item{trial}{The column from \code{tab} that is used as the trial identifier.
By default this is \code{'trial'}.}
}
\value{
A tracks object.
}
\description{
Pick a new frame as the starting frame, which will become frame 1. You can
use either a single frame number for all trials using \code{value}, or you
can supply a different frame number for each trial using \code{tab}.
}
\details{
All values below the start frame will be filtered out.
}
\examples{
  Guppies <- as_tracks(guppies, 30, '1080p')
  # Normally this dataset starts from frame 10000:
  Guppies$tr
  # But we can easily set that to 1 instead:
  set_start_frame(Guppies, 10000)$tr
  # Or we can use different frame numbers per trial:
  frame_table <- data.frame(trial = c('a', 'b'), frame = c(5000, 10000))
  set_start_frame(Guppies, tab = frame_table)$tr
  # Since some frames were fropper,

}

