% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tracks.R
\name{shiny_tracks}
\alias{shiny_tracks}
\title{Animate tracks obejct in shiny app.}
\usage{
shiny_tracks(tracks, min = NULL, max = NULL, ...)
}
\arguments{
\item{tracks}{A tracks object.}

\item{min}{Minimum range of frame.}

\item{max}{Maximum range of frame.}

\item{...}{Other parameters to be passed on to \code{\link[shiny]{runApp}}}
}
\description{
Animate tracks obejct in shiny app.
}

