% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWrapper.R
\docType{methods}
\name{show,PlotFeatureSet-method}
\alias{show,PlotFeatureSet-method}
\title{Display an object of class/superclass PlotFeatureSet}
\usage{
\S4method{show}{PlotFeatureSet}(object)
}
\arguments{
\item{object}{An object of (super)class PlotFeatureSet.}
}
