% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{simplerTracker}
\alias{simplerTracker}
\title{A Simple Tracker}
\usage{
simplerTracker(current, past, maxDist = 10)
}
\arguments{
\item{current}{A data frame with at least 4 columns: x, y, frame, and track.}

\item{past}{A data frame with at least 4 columns: x, y, frame, and track.}

\item{maxDist}{The maximum distance between two successive positions
belonging to the same trajectory.}
}
\value{
A data frame with at least 4 columns: x, y, frame, and track.
}
\description{
Given a set of current x/y positions, this function attempts to
 find which trajectory they belong to in a set of past tracked positions
 using the Hungarian method.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
