% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{sm}
\alias{sm}
\title{Smooth a vector using a rolling mean.}
\usage{
sm(x, window = 5, weights = stats::dnorm(seq(-3, 3, length.out = 9)))
}
\arguments{
\item{x}{A vector to smooth.}

\item{window}{The size of the window to smooth over (ignored if weights is
non-\code{NULL}).}

\item{weights}{A vector of weights.}
}
\value{
A numeric vector.
}
\description{
By default, smoothing is done by averaging over a Gaussian kernel.
}

