% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_tr.R
\name{speed}
\alias{speed}
\title{Calculate speed.}
\usage{
speed(x, y, f)
}
\arguments{
\item{x}{X coordinates.}

\item{y}{Y coordinates.}

\item{f}{Always use 'f' or 'frame' (equivalent). Used to order by frame
before calculation.}
}
\value{
A vector of accelerations in px per frame.
}
\description{
Calculate speed.
}

