% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R, R/accessor-methods.R
\docType{methods}
\name{statTransform<-}
\alias{statTransform}
\alias{statTransform,GGplotFeatureSet-method}
\alias{statTransform,PlotFeatureSet-method}
\alias{statTransform,gTree-method}
\alias{statTransform,ggplot-method}
\alias{statTransform,trellis-method}
\alias{statTransform<-}
\alias{statTransform<-,GGplotFeatureSet-method}
\title{Mutator method for GGplotFeatureSet "stat" slot.}
\usage{
statTransform(object) <- value

statTransform(object)

\S4method{statTransform}{PlotFeatureSet}(object)

\S4method{statTransform}{ggplot}(object)

\S4method{statTransform}{trellis}(object)

\S4method{statTransform}{gTree}(object)

\S4method{statTransform}{GGplotFeatureSet}(object) <- value

\S4method{statTransform}{GGplotFeatureSet}(object)
}
\arguments{
\item{object}{An object of class GGplotFeatureSet. For accession, also plot objects of class ggplot.}

\item{value}{A named list of parameters for statistical transforms in each layer of the plot.}
}
\value{
A modified object of class GGplotFeatureSet.

A named list of parameters for statistical transforms in each layer of the plot.
}

