% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_base_methods.R
\name{[[.tracks}
\alias{[[.tracks}
\title{Selecting a table from a tracks object.}
\usage{
\method{[[}{tracks}(tracks, table, collect = TRUE)
}
\arguments{
\item{tracks}{A tracks object.}

\item{table}{The name of a table.}

\item{collect}{Whether the table should be collected from the cluster, if
applicable.}
}
\value{
A \code{tbl_df} or a \code{party_df}.
}
\description{
Selecting a table from a tracks object.
}
\examples{
Guppies <- as_tracks(guppies, 30, 1080)

# Select and collect:
Guppies[['tr']]

# Only select:
Guppies[['tr', FALSE]]
Guppies$tr
}

