% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sum_na}
\alias{sum_na}
\title{Calculate mean of a vector or return NA if there are no non-missing values to calculate from
Useful for summarising}
\usage{
sum_na(x)
}
\arguments{
\item{x}{vector}
}
\value{

}
\description{
Calculate mean of a vector or return NA if there are no non-missing values to calculate from
Useful for summarising
}
