% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{summarise}
\alias{summarise}
\alias{summarise_}
\alias{summarise_.tracks}
\alias{summarize}
\alias{summarize_}
\title{Summarize a tracks variable into higher tables}
\usage{
\method{summarise_}{tracks}(.data, ..., .source = "tr", .tables = c("group",
  "animal", "trial"), .dots)
}
\arguments{
\item{.data}{A tracks object.}

\item{...}{Summary expressions, such as \code{mean(speed, na.rm = TRUE)}, or
\code{var_nnd = var(nearest_neighbour_dist, na.rm = TRUE)}.}

\item{.source}{Which table those variables come from. Will make guesses in
the future.}

\item{.tables}{To which tables you want to summarize them. Will make guesses
in the future.}

\item{.dots}{Used to work around non-standard evaluation. See
\code{vignette("nse")} for details.}
}
\value{
A tracks object.
}
\description{
Summarize a tracks variable into higher tables
}
\examples{
  tr <- add_speed(guppies)
  summarize(guppies, speed, .tables = 'trial')
}
\seealso{
dplyr::summarise
}

