% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-dplyr_verbs.R
\name{summarise_sections}
\alias{summarise_sections}
\alias{summarise_sections_}
\alias{summarize_sections}
\alias{summarize_sections_}
\title{Summarize track sections.}
\usage{
summarise_sections(sections, ...)

summarize_sections(sections, ...)

summarize_sections_(sections, ..., .dots)

summarise_sections_(sections, ..., .dots)
}
\arguments{
\item{sections}{Ouput from \code{find_sections} (a list of length 2).}

\item{...}{Summary statements.}

\item{.dots}{Used to work around non-standard evaluation. See vignette("nse")
for details.}
}
\value{
A tbl_df.
}
\description{
Process the output of \code{find_sections} by summarizing each tracks
portion, and joing that to the sections table.
}
\details{
Multiple rows per section can be created, as grouping is preserved in the
summary. By default, \code{tr} is grouped by animal and \code{soc} is grouped
by pair. Adjust grouping beforehand if other results are required.
}
\section{Functions}{
\itemize{
\item \code{summarize_sections}: Retrieve the timestamps for track section
based on conditions.

\item \code{summarize_sections_}: Retrieve the timestamps for track section
based on conditions.

\item \code{summarise_sections_}: Retrieve the timestamps for track section
based on conditions.
}}

