% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWrapper.R
\docType{methods}
\name{summary,PlotFeatureSet-method}
\alias{summary,PlotFeatureSet-method}
\title{Summarize an object of class/superclass PlotFeatureSet}
\usage{
\S4method{summary}{PlotFeatureSet}(object)
}
\arguments{
\item{object}{An object of (super)class PlotFeatureSet.}
}
\description{
Summary methods for PlotFeatureSet objects
}
