% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_plots.R
\name{time_facet_plot}
\alias{time_facet_plot}
\title{Facet paths using time bins.}
\usage{
time_facet_plot(tracks, x = ~X, y = ~Y, time_bins = 4, color = ~animal,
  mode = "dual", nrow = NULL, scales = "free_x", facet = trial ~
  time_bin, coord_boundary = NULL)
}
\arguments{
\item{tracks}{A tracks object.}

\item{x}{Formula for what should be on the x-axis.}

\item{y}{Formula for what should be on the y-axis.}

\item{time_bins}{The number of periods time should be divided in.}

\item{color}{Formula to color the paths.}

\item{mode}{Either dual, which uses both the rows and columns of the facets
to display time bins, or manual, where you can manually set the formula for
facet_grid.}

\item{nrow}{Override the number of rows that should be used in facetting.
Only used if mode is dual.}

\item{facet}{Optional facetting, should probably include time_bin. Only used
if mode is set to manual.}

\item{coord_boundary}{Whether to fix the plot limits to the boundary. Will
attempt to autodetect whether plotting ~X and ~Y.}
}
\value{
A ggplot object.
}
\description{
Facet paths using time bins.
}
\examples{
  time_facet_plot(filter(guppies, trial == 'd1t3b'))
  time_facet_plot(guppies, mode = 'manual')
}

