% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{history_tracker}
\alias{history_tracker}
\alias{knitr_tracker}
\title{Construct and activate a HistoryTracker object}
\usage{
history_tracker(id = "history_tracker")

knitr_tracker()
}
\arguments{
\item{id}{character. The id of the taskback}
}
\value{
A HistoryTracker object
}

