% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotrack.R
\name{trackHistory}
\alias{trackHistory}
\title{Automatically track history within an R session}
\usage{
trackHistory(tracker = NULL)
}
\arguments{
\item{tracker}{VirtHistoryTracker subclass or NULL. For NULL, if
a default tracker is set, toggle tracking with default tracker, otherwise
set a default tracker. For a VirtHistoryTracker (subclass) set as
the default tracker and turn it on.}
}

