% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr-package.R
\docType{package}
\name{trackr-package}
\alias{trackr}
\alias{trackr-package}
\title{trackr: Track Progress and Forecast Targets for WDI Indicators}
\description{
Calculates progress scores and future targets for any WDI indicator of interest. Users can specify their preferred method and parameters to evaluate and project country performance. Depending on the settings, the package returns between two and six data frames summarizing percentile-based and speed-based scores, as well as historical and future indicator trajectories under different assumptions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RossanaTat/sdgtrackr}
  \item Report bugs at \url{https://github.com/RossanaTat/sdgtrackr/issues}
}

}
\keyword{internal}
