% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackR.R
\name{trackR}
\alias{trackR}
\title{Simple Video Tracking Software}
\usage{
trackR(tracker = "classic", ...)
}
\arguments{
\item{tracker}{A character string indicating the tracking algorithm to use
(see 'Details').}

\item{...}{Arguments passed to \link[shiny]{runApp}.}
}
\value{
This function does not return anything. A file is saved if the
 tracking is successfully completed.
}
\description{
Simple video tracking software with Shiny-based GUI.
}
\details{
TODO
}
\examples{
# Run trackR in an external window
\dontrun{trackR()}

# Run trackR inside RStudio
\dontrun{trackR(launch.browser = rstudioapi::viewer)}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
