% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjFeatureSet.R
\docType{data}
\name{trackr_knit_env}
\alias{trackr_knit_env}
\title{Internal details}
\format{An object of class \code{environment} of length 0.}
\usage{
trackr_knit_env
}
\description{
This environment should never be manipulated or
    interacted with directly by end users. It is exported only due
    the vagaries of how parts of trackr are implemented.
}
\keyword{datasets}
