% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{trackrAddin}
\alias{trackrAddin}
\title{RStudio addin/Shiny app for artifact discovery}
\usage{
trackrAddin(tdb = defaultTDB())
}
\arguments{
\item{tdb}{The database to search. Defaults to the current default database}
}
\description{
This function initiates a shiny server which  allows users
to search a trackr database and view the results. It can
be used as a stand-alone Shiny application or from within the
RStudio IDE as an addin
}
