% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{TrackrDB-class}
\alias{TrackrDB}
\alias{TrackrDB-class}
\title{Trackr database}
\usage{
TrackrDB(opts = TrackrOptions(...), backend = JSONBackend(), ...)
}
\arguments{
\item{opts}{TrackrOptions object.}

\item{backend}{ANY. The backend to use.}

\item{...}{ignored.}
}

