% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry.R
\name{trig}
\alias{angle}
\alias{angle_diff}
\alias{mean_angle}
\alias{trig}
\title{Common functions to deal with angles.}
\usage{
angle(x1, y1, x2, y2)

angle_diff(a1, a2)

mean_angle(x)
}
\arguments{
\item{x1}{X-coordinate for poinr 1.}

\item{y1}{Y-coordinate for poinr 1.}

\item{x2}{X-coordinate for poinr 2.}

\item{y2}{Y-coordinate for poinr 2.}

\item{a1}{Angle 1.}

\item{a2}{Angle 2.}

\item{x}{A vector of angles.}
}
\description{
These functions make it easy to calculate things about orientation and
headings, while keeping the angles within -pi to pi.
}

