% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_tr.R
\name{turn}
\alias{turn}
\title{Calculate turning angle.}
\usage{
turn(x, y, f)
}
\arguments{
\item{x}{X coordinates.}

\item{y}{Y coordinates.}

\item{f}{Always use 'f' or 'frame' (equivalent). Used to order by frame
before calculation.}
}
\description{
This adds the turning angle of the path, that is, based on coordinates, not
orientation.
}

