% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackR.R
\name{videoFixer}
\alias{videoFixer}
\title{Fix Video Recording Errors}
\usage{
videoFixer(...)
}
\arguments{
\item{...}{Arguments passed to \link[shiny]{runApp}.}
}
\value{
This function does not return anything. A file is saved if the
 export button is used in the app.
}
\description{
Simple Shiny-based GUI to fix certain common video recording
 errors in behavioral experiment, such as frame shifting, and color and light
 variations.
}
\examples{
# Run videoFixer in an external window
\dontrun{videoFixer()}

# Run videoFixer classic inside RStudio
\dontrun{videoFixer(launch.browser = rstudioapi::viewer)}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
