% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackR.R
\name{videoMaker}
\alias{videoMaker}
\title{Convert Image Sequence to Video}
\usage{
videoMaker(...)
}
\arguments{
\item{...}{Arguments passed to \link[shiny]{runApp}.}
}
\value{
This function does not return anything. A file is saved if the
 export button is used in the app.
}
\description{
Simple Shiny-based GUI to convert a sequence of images into a
 video compatible with \code{\link{trackR}}.
}
\examples{
# Run videoMaker in an external window
\dontrun{videoMaker()}

# Run videoMaker classic inside RStudio
\dontrun{videoMaker(launch.browser = rstudioapi::viewer)}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
