% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{ViztrackrDB-class}
\alias{ViztrackrDB}
\alias{ViztrackrDB-class}
\title{Viztrackr database}
\usage{
ViztrackrDB(opts = ViztrackrOptions(...), backend = VTJSONBackend(), ...)
}
\arguments{
\item{opts}{ViztrackrOptions object.}

\item{backend}{ANY. The backend to use.}

\item{...}{ignored.}
}

