% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{VTJSONBackend}
\alias{VTJSONBackend}
\title{JSON backend contstructor}
\usage{
VTJSONBackend(file = normalizePath("./viztrackr_db_data.json"),
  data = list())
}
\arguments{
\item{file}{character. The json "database" to use as a viztrackr backend}
}
\value{
A VTJSONBackend object, for use in creating a ViztrackrDB object.
}

