% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotrack.R
\docType{data}
\name{vtopts}
\alias{vtopts}
\title{viztrackr options environment
An environment where the default vtdb and history tracker are stored.
Users should not modify the contents of this environment directly.}
\format{An object of class \code{environment} of length 3.}
\usage{
vtopts
}
\keyword{datasets}

