% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_pairs.R
\name{add_pair_nip_dist_velocity}
\alias{add_pair_nip_dist_velocity}
\title{Title}
\usage{
add_pair_nip_dist_velocity(tracks)
}
\arguments{
\item{tracks}{A tracks object.}
}
\value{
A tracks objec
}
\description{
Title
}
\seealso{
add_pair_nip_dist
}

