% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjFeatureSet.R
\name{as.list}
\alias{as.list}
\alias{as.list.FeatureSet}
\alias{as.list,FeatureSet-method}
\alias{as.list.ScriptNodeInfo}
\title{Convert objects to lists}
\usage{
\method{as.list}{FeatureSet}(x, ...)

\method{as.list}{ScriptNodeInfo}(x, ...)
}
\arguments{
\item{x}{The object to convert to a list.}

\item{...}{Other named arguments (currently unused).}
}
\value{
A list.
}
\description{
Convert objects to lists.
}
