% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotrack.R
\name{autotrackPlots}
\alias{autotrackPlots}
\title{Automatically record plots
Toggle automatic recording of plots to the
current default ViztrackrDB.}
\usage{
autotrackPlots(vtdb = vtopts$vtdb)
}
\arguments{
\item{vtdb}{ViztrackrDB. The ViztrackrDB to record
plots to automatically. If specified, will become
the default.}
}
\value{
NULL, invisibly.
}

