% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/accessor-methods.R
\docType{methods}
\name{codeInfo<-}
\alias{codeInfo}
\alias{codeInfo,ObjFeatureSet-method}
\alias{codeInfo<-}
\alias{codeInfo<-,ObjFeatureSet-method}
\title{Mutator method for "code.info" slot of PlotFeatureSet (or one of its subclasses).}
\usage{
codeInfo(object) <- value

codeInfo(object)

\S4method{codeInfo}{ObjFeatureSet}(object) <- value

\S4method{codeInfo}{ObjFeatureSet}(object)
}
\arguments{
\item{object}{An object of (super)class PlotFeatureSet. For accession, also plot objects of class ggplot or trellis.}

\item{value}{A CodeDepends::ScriptInfo object.}
}
\value{
A modified object of (super)class PlotFeatureSet.

A CodeDepends::ScriptInfo object
}

