% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/accessor-methods.R
\docType{methods}
\name{codeInfo<-}
\alias{codeInfo<-}
\alias{codeInfo}
\alias{codeInfo<-,FeatureSet-method}
\alias{codeInfo,FeatureSet-method}
\title{Accessors for the code analysis information of FeatureSet objects}
\usage{
codeInfo(object) <- value

codeInfo(object)

\S4method{codeInfo}{FeatureSet}(object) <- value

\S4method{codeInfo}{FeatureSet}(object)
}
\arguments{
\item{object}{An object of (super)class PlotFeatureSet. For accession, also plot objects of class ggplot or trellis.}

\item{value}{A CodeDepends::ScriptInfo object.}
}
\value{
A modified object of (super)class PlotFeatureSet.
}
\description{
Get or set the analysis information for the code associated with a FeatureSet object. This should not be called directly.
}
\examples{
fs = makeFeatureSet(mtcars)
codeInfo(fs)
}
