% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_constructors.R
\name{jsonVTDB}
\alias{jsonVTDB}
\alias{solrVTDB}
\title{Convenience constructors for specific VT backends}
\usage{
jsonVTDB(file, opts = ViztrackrOptions(...), ...)

solrVTDB(core, requestHandler = "search", opts = ViztrackrOptions(...), ...)
}
\arguments{
\item{file}{character. The json file to use}

\item{opts}{ViztrackrOptions. The options for the DB}

\item{core}{character. The URI for the solr core to use}

\item{requestHandler}{character. Passed to \code{SolrList} constructor}

\item{\dots{}}{ANY. Used to construct \code{opts}. Ignored if
\code{opts} is specified explicitly.}
}
\value{
A ViztrackrDB object
}

