% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R
\docType{methods}
\name{describePackage}
\alias{describePackage}
\alias{describePackage,FeatureSet-method}
\title{Accessor method for ObjFeatureSet "analysispkg" slot.}
\usage{
describePackage(object)

\S4method{describePackage}{FeatureSet}(object)
}
\arguments{
\item{object}{An object of (super)class ObjFeatureSet.}
}
\value{
A list holding the name, title, and description of any
    packages with description files on the current path.
}
\description{
The analysis package, if any, that was 'in use' (i.e.,
    that contained the working directory that was active) when the
    result was recorded.
}
\examples{
mt = datasets::mtcars
fs = makeFeatureSet(mt)
describePackage(fs)
}
