% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{methods}
\name{DocCollectionRef}
\alias{DocCollectionRef}
\alias{.docrefclass}
\alias{DocCollectionRef-class}
\alias{names<-,DocCollectionRef,ANY-method}
\alias{ndoc,DocCollectionRef-method}
\alias{ListBackend-class}
\alias{listbackend}
\alias{ListBackend}
\title{Reference to a DocCollection}
\usage{
ListBackend(lst = list())
}
\arguments{
\item{lst}{List of documents to populate the list backend with.}
}
\description{
A reference class which carries around a
    DocCollection, suitable for use as a trackr backend.
}

