% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena.R
\name{estimate_arena}
\alias{estimate_arena}
\title{Estimate arena from tracks and reallign.}
\usage{
estimate_arena(tracks, grouping = ~trial, shape = "circle", radius = NULL)
}
\arguments{
\item{tracks}{A tracks object.}

\item{grouping}{On what level the arena should be estimated.}

\item{shape}{The shape of the arena. Currently has to be a circle.}

\item{radius}{The true radius of the circle, used for rescaling the track
coordinates.}
}
\value{
A tracks object.
}
\description{
Estimate arena from tracks and reallign.
}

