% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-dplyr_verbs.R
\name{find_sections}
\alias{find_sections}
\title{Retrieve the timestamps for track sections based on conditions.}
\usage{
find_sections(tracks, ..., tol = 0)
}
\arguments{
\item{tracks}{A tracks object.}

\item{...}{Conditions.}

\item{tol}{Combine sequences that are \code{tol} frames apart.}

\item{.dots}{Used to work around non-standard evaluation. See vignette("nse")
for details.}
}
\value{
A tbl_df.
}
\description{
Similar to \code{filter}, but returns a table of sequences grouped by trial,
with a \code{start}, \code{stop} and \code{length} column. It supports a
tolerance level which allows for combining sequences that are close together
(the default of 1 is zero tolerance).
}
\details{
Any variables used as conditions will be looked up in the \code{tr},
\code{soc} and \code{group} tables and applied when present. Non-existing
variables will not produce an error.

Seperate conditions on different variables with a \code{,}, not \code{&}.
Combine several conditions on the same variable with \code{&}.

Example use case: If one wants to select all sequences of frames where
animals are chasing each other (to plot them, for example), one could filter
for a high mean speed of the pair, and a small pairwise distance.
}

