% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry.R
\name{find_smallest_enclosing_circle}
\alias{find_smallest_enclosing_circle}
\title{Find smalles enclosing circle.}
\usage{
find_smallest_enclosing_circle(points)
}
\arguments{
\item{points}{A matrix or data.frame with the first column being x and the
second being y.}
}
\value{
A vector of lenght three, center x and y, and radius.
}
\description{
Find smalles enclosing circle.
}

