% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_paths.R
\name{future_path}
\alias{future_path}
\title{Wrapper to Compute Future Paths Based on Either Percentiles or Speeds method}
\usage{
future_path(
  data_fut,
  target_year = 2030,
  min = 0,
  max = 100,
  granularity = 0.1,
  sequence_pctl = seq(20, 80, 20),
  changes_pctl = NULL,
  sequence_speed = c(0.25, 0.5, 1, 2, 4),
  path_speed = NULL,
  best = "high",
  verbose = TRUE,
  speed = FALSE,
  percentiles = FALSE
)
}
\arguments{
\item{data_fut}{A data frame containing historical data with columns `code`, `year`, and `y`. This serves as the baseline for projections.}

\item{target_year}{An integer specifying the year to project to. Default is `2030`.}

\item{min}{Optional. Minimum value of `initialvalue` to predict.}

\item{max}{Optional. Maximum value of `initialvalue` to predict.}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.}

\item{sequence_pctl}{Percentiles to calculate. The percentile score assigned is determined by the granularity of the percentiles selected here. I.e., if seq(20,80,20) is selected, then we will be able to tell if a country is between 0-20, 20-40, etc.}

\item{changes_pctl}{A `data.table` with predicted changes by initialvalue and pctl.}

\item{best}{A character string indicating the direction of improvement.
Use \code{"high"} (default) if higher values of the indicator are better,
or \code{"low"} if lower values are better. This affects the direction of the path.}

\item{verbose}{Logical; if `TRUE`, messages will be printed for each processing year. Default is `TRUE`.}

\item{speed}{Logical. If `TRUE`, calls `future_path_speed()`.}

\item{percentiles}{Logical. If `TRUE`, calls `future_path_pctls()`.}
}
\value{
A `data.table` with projected paths based on selected method.
}
\description{
Computes future trajectories for each country either based on percentile growth projections
(`future_path_pctls()`) or speed of progress (`future_path_speed()`), depending on user input.
}
\seealso{
[future_path_pctls()], [future_path_speed()]
}
