% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_paths.R
\name{future_path_pctls}
\alias{future_path_pctls}
\title{Generate Future Percentile Paths}
\usage{
future_path_pctls(
  data_fut,
  sequence_pctl = seq(20, 80, 20),
  predictions_pctl,
  target_year = 2030,
  granularity = 0.1,
  min = 0,
  max = 100,
  verbose = TRUE
)
}
\arguments{
\item{data_fut}{A data frame containing historical data with columns `code`, `year`, and `y`. This serves as the baseline for projections.}

\item{sequence_pctl}{Percentiles to calculate. The percentile score assigned is determined by the granularity of the percentiles selected here. I.e., if seq(20,80,20) is selected, then we will be able to tell if a country is between 0-20, 20-40, etc.}

\item{predictions_pctl}{A `data.table` with predicted changes by initialvalue and pctl.}

\item{target_year}{An integer specifying the year to project to. Default is `2030`.}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.}

\item{min}{Optional. Minimum value of `initialvalue` to predict. If `NULL`,
it is set to `floor` (if provided) or `min(data_model$initialvalue)`.}

\item{max}{Optional. Maximum value of `initialvalue` to predict. If `NULL`,
it is set to `ceiling` (if provided) or `max(data_model$initialvalue)`.}

\item{verbose}{Logical; if `TRUE`, messages will be printed for each processing year. Default is `TRUE`.}
}
\value{
A data.table with the projected values (`y_fut`) by `code`, `year`, and `pctl`, including observed values where available.
}
\description{
Computes projected future paths for given percentiles based on initial levels and predicted changes.
}
\details{
This function builds a dataset that extends observed data (`data_fut`) forward in time to a target year using
predicted changes at specified percentiles. It iteratively calculates future values by applying changes year-by-year,
starting from the last available observation for each country-percentile pair.
}
\examples{
\dontrun{
data_fut <- data.frame(code = c("A", "A"), year = c(2020, 2021), y = c(50, 52))
predictions_pctl <- data.frame(initialvalue = 52, pctl = 20, change = 1)
future_path_pctls(data_fut = data_fut, predictions_pctl = predictions_pctl, target_year = 2025)
}

}
