% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_paths.R
\name{future_path_speed}
\alias{future_path_speed}
\title{Project future indicator paths at different speeds}
\usage{
future_path_speed(
  data_fut,
  sequence_speed = c(0.25, 0.5, 1, 2, 4),
  path_speed,
  best = "high",
  target_year = 2030,
  min = NULL,
  max = NULL
)
}
\arguments{
\item{data_fut}{A data frame or data table containing the historical and/or
baseline data. Must include columns: `code`, `year`, and `y` (the indicator),
and optionally `y_fut` and `time` for projections.}

\item{best}{A character string indicating whether higher or lower values of thei indicator are better. Use "high" if higher values of the indicator reflect better outcomes, or "low" if lower values are better.}

\item{target_year}{An integer specifying the year to project to. Default is '2030'. Only relevant if future=TRUE.}

\item{min}{Optional. Minimum value of `initialvalue` to predict.}

\item{max}{Optional. Maximum value of `initialvalue` to predict.}
}
\value{
A data table with projected indicator values (`y_fut`) for each
  country-year-speed combination from the last available year through
  the `target_year`, respecting the selected progress speeds and
  indicator direction (`best`).
}
\description{
This function simulates future indicator trajectories under different speeds of progress. It generates yearly projections from the latest available data point up to a given target year.
}
