% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_paths.R
\name{get_his_data}
\alias{get_his_data}
\title{Get historical baseline data}
\usage{
get_his_data(
  indicator = "EG.ELC.ACCS.ZS",
  data = NULL,
  code_col = "iso3c",
  year_col = "date",
  min = NULL,
  max = NULL,
  eval_from = 2000,
  eval_to = 2022,
  support = 1,
  granularity = 0.1
)
}
\arguments{
\item{indicator}{Character. Name of the column with the values of the indicator data}

\item{data}{Optional. A data frame with indicator data.}

\item{code_col}{Character. Name of the column with country codes (or other groupings used to calculate and compare progress)}

\item{year_col}{Character. Name of the column with years.}

\item{min}{Numeric (optional). Minimum value of the indicator. If NULL, the minimum is computed from the data and rounded to the nearest multiple of `granularity`.
Will be overruled by support if that reflects a higher minimum value}

\item{max}{Numeric (optional). Maximum value of the indicator. If NULL, the maximum is computed from the data and rounded to the nearest multiple of `granularity`. Will be overruled by support if that reflects a lower maximum value.}

\item{eval_from}{First year to include when evaluating the progress of countries.}

\item{eval_to}{Last year to include when evaluating the progress of countries.}

\item{support}{Numeric. Reflects minimum number of countries that must have experienced a particular outcome value from startyear_data to endyear_data for a progress score to be calculated. This limits the calculation of scores at extreme outcome values at which there is less support to evaluate what typical progress looks like. Will be overruled by min and max if those are more restrictive. Defaults to 1.}

\item{granularity}{Numeric. Granularity in outcome variable. All outcome values are rounded to the nearest value reflecting this granularity. For example, if 1 is chosen, then all outcome values are rounded to the nearest integer. Default is 0.1. The lower granularity, the more precise the estimated paths but the longer it takes to run the code.}
}
\value{
A `data.table` containing the following columns:
  \item{code}{Country code (standardized).}
  \item{year}{Year of observation.}
  \item{y}{Value of the indicator.}
  \item{y_his}{First non-missing historical value (rounded to the specified granularity).}
}
\description{
Extract and format historical indicator data for each country to establish a baseline (y_his)
used to evaluate progress. This function selects data within a given time range,
filters out leading missing values, and stores the first valid (non-missing) observation—
rounded to a specified granularity—as the baseline value (`y_his`).
}
\examples{
\dontrun{
  his_data <- get_his_data(indicator = "EG.ELC.ACCS.ZS",
                           eval_from = 2000,
                           eval_to = 2022,
                           granularity = 0.1)
}

}
