% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_paths.R
\name{get_his_data}
\alias{get_his_data}
\title{Get historical baseline data}
\usage{
get_his_data(
  indicator = "EG.ELC.ACCS.ZS",
  data = wbstats::wb_data(indicator = indicator, lang = "en", country = "countries_only"),
  code_col = "iso3c",
  year_col = "date",
  min = 0,
  max = 100,
  eval_from = 2000,
  eval_to = 2022,
  granularity = 0.1
)
}
\arguments{
\item{indicator}{Character. Indicator code (e.g., `"EG.ELC.ACCS.ZS"`). Defaults to access to electricity.}

\item{data}{Optional. A data frame with indicator data. If NULL, data is downloaded via `wbstats::wb_data()`.}

\item{code_col}{Character. Name of the column with country codes. Defaults to `"iso3c"`.}

\item{year_col}{Character. Name of the column with years. Defaults to `"date"`.}

\item{min}{A numeric value indicating the minimum bound,
typically passed from a previously processed dataset (e.g., `prep_data()` output via `min <- data_model$min`).
This value is stored as an attribute of the returned object for later reference (e.g., in visualizations or simulations).}

\item{max}{A numeric value indicating the maximum bound,
typically passed from a previously processed dataset (e.g., `prep_data()` output via `max <- data_model$max`).
This value is stored as an attribute of the returned object for later reference (e.g., in visualizations or simulations).}

\item{eval_from}{The first year to include in the analysis}

\item{eval_to}{The last year to include in the analysis}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.}
}
\value{
A `data.table` containing the following columns:
  \item{code}{Country code (standardized).}
  \item{year}{Year of observation.}
  \item{y}{Value of the indicator.}
  \item{y_his}{First non-missing historical value (rounded to the specified granularity).}
}
\description{
Extract and format historical indicator data for each country to establish a baseline (y_his)
used to evaluate progress. This function selects data within a given time range,
filters out leading missing values, and stores the first valid (non-missing) observation—
rounded to a specified granularity—as the baseline value (`y_his`).
}
\examples{
\dontrun{
  his_data <- get_his_data(indicator = "EG.ELC.ACCS.ZS",
                           eval_from = 2000,
                           eval_to = 2022,
                           granularity = 0.1)
}

}
