% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_paths.R
\name{get_his_data}
\alias{get_his_data}
\title{Get historical baseline data}
\usage{
get_his_data(
  indicator = "EG.ELC.ACCS.ZS",
  data = wbstats::wb_data(indicator = indicator, lang = "en", country = "countries_only"),
  code_col = "iso3c",
  year_col = "date",
  start_year = 2000,
  end_year = 2022,
  granularity
)
}
\arguments{
\item{indicator}{Character. Indicator code or name (e.g., `"EG.ELC.ACCS.ZS"`). Defaults to access to electricity.}

\item{data}{Optional. A data frame with indicator data. If NULL, data is downloaded via `wbstats::wb_data()`.}

\item{code_col}{Character. Name of the column with country codes. Defaults to `"iso3c"`.}

\item{year_col}{Character. Name of the column with years. Defaults to `"date"`.}

\item{granularity}{Numeric. Step size for the prediction sequence. Default is `0.1`.}
}
\value{
A `data.table` containing the following columns:
  \item{code}{Country code (standardized).}
  \item{year}{Year of observation.}
  \item{y}{Value of the indicator.}
  \item{y_his}{First non-missing historical value (rounded to the specified granularity).}
}
\description{
Extract and format historical indicator data for each country to establish a baseline (y_his)
used to evaluate progress. This function selects data within a given time range,
filters out leading missing values, and stores the first valid (non-missing) observation—
rounded to a specified granularity—as the baseline value (`y_his`).
}
\examples{
\dontrun{
  his_data <- get_his_data(indicator = "EG.ELC.ACCS.ZS",
                           start_year = 2000,
                           end_year = 2022,
                           granularity = 0.1)
}

}
