% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{get_scores_speed}
\alias{get_scores_speed}
\title{Calculate speed-based score over time}
\usage{
get_scores_speed(
  path_his_speed,
  path_speed,
  min = NULL,
  max = NULL,
  granularity = 0.1
)
}
\arguments{
\item{path_his_speed}{Result of `project_pctls_path()`}

\item{path_speed}{Result of `get_speed_path()`}

\item{min}{Numeric (optional). Minimum value of the indicator. If NULL, the minimum is computed from the data and rounded to the nearest multiple of `granularity`.}

\item{max}{Numeric (optional). Maximum value of the indicator. If NULL, the maximum is computed from the data and rounded to the nearest multiple of `granularity`.}

\item{granularity}{Rounding precision for `y`. Default is 0.1.}
}
\value{
A data.table with `code`, `score`, and `evaluationperiod`.
}
\description{
Computes a score based on the average annual change in percentile rank over at least 5 years.
}
