% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-methods.R
\docType{methods}
\name{trackr_options}
\alias{trackr_options}
\alias{img_dir}
\alias{img_ext}
\title{options extraction}
\usage{
trackr_options(db)

img_dir(db)

img_ext(db)
}
\arguments{
\item{db}{Object to extract information from}
}
