% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\docType{methods}
\name{vt_lookup}
\alias{insert_plot}
\alias{prep_for_vtdb}
\alias{remove_plot}
\alias{vt_grep}
\alias{vt_lookup}
\title{Backend-interface API

This page describes the API which must be implimented by all Viztrackr backends
Any class with methods for these generic can be used in the \code{backend}
slot of a \code{ViztrackrDB} object.

Methods should be written to be dispatched \code{backend}, and, where appropriate
on code{object}/code{doc}}
\usage{
vt_lookup(object, target, opts, exist = FALSE)

insert_plot(object, id, target, opts, verbose = FALSE)

prep_for_vtdb(object, target, opts, verbose = FALSE)

remove_plot(object, target, opts, verbose = FALSE)

vt_write(target, opts, verbose = FALSE)

vt_grep(pattern, target, opts, fields = NULL, ret_type = c("doclist", "id",
  "backend"), verbose = TRUE)
}
\arguments{
\item{object}{ANY. The object to lookup, add, remove, etc.}

\item{target}{ANY. The backend of the ViztrackrDB instance.}

\item{opts}{ViztrackrOptions. The viztrackr-level options. Typically
extracted from \code{target} in a \code{ViztrackrDB} method and passed down.}

\item{exist}{logical. Return TRUE/FALSE rather than the looked-up object.
(default: FALSE)}

\item{verbose}{logical. Should extra informative messages be displayed (
default: FALSE)}

\item{pattern}{character. A regular expression to match against the text in \code{fields}}

\item{fields}{character or NULL. The fiends in which to match, or NULL to
include all fields.}

\item{ret_type}{character. Format in which to return the response. Options are:
"id" - id of matching documents (default), "doclist" - A list containing the
matching documents represnted as R lists, and "backend" - a backend specific
representation of the set of matching documents (generally the same class
as \code{backend}.}
}
\value{
for \code{vt_lookup}: If \code{exist} is TRUE, a logical indicating whether \code{object} was
found in \code{db}. Otherwise, the object stored in the database (or NULL if it
was not found).

for \code{insert_plot} and \code{remove_plot}: The \code{ViztrackrDB}
(\code{db} parameter) after the plot has been added or removed.

For \code{prep_for_vtdb}, \code{object}, representend in the form
that the \code{insert_plot} method for \code{backend} expects.
}
\details{
\code{insert_plot} may or may not involve writing to
disk, which can alternatively occur during \code{vt_write}. Writing,
if any is desired, must occur within at one and only one of these methods. If
\code{insert_plot} performs the writing, \code{vt_write} should be a no-op.
}
\note{
\code{remove_plot} should have the same writing behavior
as \code{insert_plot}

\code{remove_plot} should have the same writing behavior
as \code{insert_plot}
}

