% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump.R
\name{jump}
\alias{jump}
\title{Compute jump distance between consecutive localisations}
\usage{
jump(data, spaceRes)
}
\arguments{
\item{data}{Dataframe with track (trace identification) and t (frame or time) variable}

\item{spaceRes}{Pixel size}
}
\description{
Extra things
}
\examples{
a <- 2
b <- 2 * a

}
\author{
Julien Godet
}
