% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-impls.R
\docType{methods}
\name{make_image_files}
\alias{make_image_files}
\alias{make_image_files,PlotFeatureSet-method}
\alias{make_image_files,RmdFeatureSet-method}
\alias{make_image_files,ObjFeatureSet-method}
\alias{make_image_files,FeatureSet-method}
\alias{make_image_files,ANY-method}
\title{Make image files for a featureset}
\usage{
make_image_files(object, opts)

\S4method{make_image_files}{PlotFeatureSet}(object, opts)

\S4method{make_image_files}{RmdFeatureSet}(object, opts)

\S4method{make_image_files}{ObjFeatureSet}(object, opts)

\S4method{make_image_files}{FeatureSet}(object, opts)

\S4method{make_image_files}{ANY}(object, opts)
}
\arguments{
\item{object}{A FeatureSet object}

\item{opts}{Options}
}
\value{
A named list with two entries: preview.path, and
    image.path. These should be paths to (now) existing iamge
    fiels for thumbnail and main display, respectively
}
\description{
Make image files for a featureset
}
