% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{mutate}
\alias{mutate}
\alias{mutate_}
\alias{mutate_.tracks}
\title{Add new variables to a tracks object.}
\usage{
\method{mutate_}{tracks}(tracks, ..., .dots)
}
\arguments{
\item{...}{Name-value pairs of expressions. Use \code{NULL} to drop
a variable.}

\item{.dots}{Used to work around non-standard evaluation. See
\code{vignette("nse")} for details.}

\item{.data}{A tracks object}
}
\value{
A tracks object.
}
\description{
Add new variables to a tracks object.
}

