% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{mutate}
\alias{mutate}
\alias{mutate_}
\alias{mutate_.tracks}
\title{Add new variables to a tracks object.}
\usage{
\method{mutate_}{tracks}(.data, ..., .dots)
}
\arguments{
\item{.data}{A tracks object}

\item{...}{Name-value pairs of expressions. Use \code{NULL} to drop
a variable.

These arguments are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in the context of the data
frame. They support \link[rlang:quasiquotation]{unquoting} and
splicing. See \code{vignette("programming")} for an introduction to
these concepts.}
}
\value{
A tracks object.
}
\description{
Add new variables to a tracks object.
}
