% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_tr.R
\name{mutate_tr}
\alias{acceleration}
\alias{angular_velocity}
\alias{change}
\alias{heading}
\alias{mutate_tr}
\alias{speed}
\title{Conveniently add common new variables to the \code{soc} table.}
\usage{
change(x, order_by = frame, frame_rate = tracks$params$frame_rate, n = 1)

speed(x = X, y = Y, order_by = frame,
  frame_rate = tracks$params$frame_rate)

acceleration(x = X, y = Y, order_by = frame,
  frame_rate = tracks$params$frame_rate)

heading(x = X, y = Y, order_by = frame)

angular_velocity(x = X, y = Y, order_by = frame,
  frame_rate = tracks$params$frame_rate)
}
\arguments{
\item{x}{X-coordinate.}

\item{order_by}{This parameter controls the ordering of \code{dplyr::lag} and
\code{dplyr::lead} and it is strongly advised to leave it at it's default
(\code{frame}).}

\item{frame_rate}{All variables with a time dimension, will be calculated
based on seconds (not frames). By default this is read from the tracks
object.}

\item{n}{Offset for change. See \code{\link[dplyr]{lag}}.}

\item{y}{Y-coordinate.}
}
\description{
These functions are designed for use within \code{summarise_}, and will
compute common derived parameters for tracks. By default, these functions
will use the expected variable names, as are default in \code{tracks}
objects, but they can be overridden.
}
\seealso{
mutate_soc mutate_.tracks
}

