% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R,
%   R/accessor-methods.R
\name{nLayers<-}
\alias{nLayers<-}
\alias{nLayers}
\alias{nLayers,PlotFeatureSet-method}
\alias{nLayers,ggplot-method}
\alias{nLayers,trellis-method}
\alias{nLayers,gTree-method}
\alias{nLayers<-,GGplotFeatureSet-method}
\alias{nLayers,GGplotFeatureSet-method}
\title{Accessors for the number of layers of GGplotFeatureSet objects}
\usage{
nLayers(object) <- value

nLayers(object)

\S4method{nLayers}{PlotFeatureSet}(object)

\S4method{nLayers}{ggplot}(object)

\S4method{nLayers}{trellis}(object)

\S4method{nLayers}{gTree}(object)

\S4method{nLayers}{GGplotFeatureSet}(object) <- value

\S4method{nLayers}{GGplotFeatureSet}(object)
}
\arguments{
\item{object}{An object of class GGplotFeatureSet. For accession, also plot objects of class ggplot.}

\item{value}{An integer representing the number of layers in the plot.}
}
\value{
A modified object of class GGplotFeatureSet.
}
\description{
Get or set the number of layers associated with a GGplotFeatureSet object. This should not be called directly.
}
\examples{
library(ggplot2)
mt = datasets::mtcars
plt = qplot(mt$mpg, mt$cyl)
fs = makeFeatureSet(plt)
nLayers(fs)
}
