% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/accessor-methods.R
\name{objCode<-}
\alias{objCode<-}
\alias{objCode}
\alias{objCode<-,FeatureSet-method}
\alias{objCode,FeatureSet-method}
\title{Accessors for the code of FeatureSet objects}
\usage{
objCode(object) <- value

objCode(object)

\S4method{objCode}{FeatureSet}(object) <- value

\S4method{objCode}{FeatureSet}(object)
}
\arguments{
\item{object}{An object of (super)class PlotFeatureSet. For accession, also plot objects of class ggplot or trellis.}

\item{value}{A CodeDepends::Script object.}
}
\value{
A modified object of (super)class PlotFeatureSet.
}
\description{
Get or set the code associated with a FeatureSet object. This should not be called directly.
}
\examples{
library(ggplot2)
mt = datasets::mtcars
plt = qplot(mt$mpg, mt$cyl)
fs = makeFeatureSet(plt)
objCode(fs)
}
