% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_paths.R
\name{path_historical}
\alias{path_historical}
\title{Compute historical paths}
\usage{
path_historical(
  percentiles = TRUE,
  speed = TRUE,
  data_his,
  eval_from = NULL,
  eval_to = NULL,
  granularity = 0.1,
  min = 0,
  max = 100,
  best = NULL,
  sequence_pctl = seq(20, 80, 20),
  changes_pctl = NULL,
  sequence_speed = c(0.25, 0.5, 1, 2, 4),
  path_speed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{percentiles}{Logical. If TRUE, calculate percentile scores. Default is TRUE}

\item{speed}{Logical. If TRUE, calculate speed of progress scores. Default is FALSE.}

\item{data_his}{A data.table containing historical values.}

\item{eval_from}{First year to include when evaluating the progress of countries.}

\item{eval_to}{Last year to include when evaluating the progress of countries.}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.}

\item{min}{Optional. Minimum value of `initialvalue` to predict.}

\item{max}{Optional. Maximum value of `initialvalue` to predict.}

\item{best}{A character string indicating whether higher or lower values of thei indicator are better. Use "high" if higher values of the indicator reflect better outcomes, or "low" if lower values are better.}

\item{sequence_pctl}{Numeric vector of percentile paths to calculate.
The percentile score is determined by the granularity of the percentiles chosen here. For example, if seq(20,80,20) is chosen, then a country’s progress will fall in the 0th-20th percentile, 20th-40th percentile etc.
This argument is only relevant if `percentiles=TRUE`.}

\item{changes_pctl}{A `data.table` with predicted changes by initialvalue and pctl.}

\item{path_speed}{Data table  with speed paths}

\item{verbose}{Logical. If TRUE, prints info messages in console}
}
\value{
A named list with one or both of `percentile_path` and `speed_path`.
}
\description{
Wrapper to compute historical paths by percentiles and/or speed
}
