% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiffusionMap.r
\name{plotDiffusionMap}
\alias{plotDiffusionMap}
\title{Plot Diffusion Map}
\usage{
plotDiffusionMap(data, traceNb, nbPix = 20, jumpMax, nbCol = 10)
}
\arguments{
\item{data}{data frame (muste contain trace, x, y, and jump)}

\item{traceNb}{(optional) number id of the traces to plot}

\item{nbPix}{number of pixel (or calibrated distance) to add on each x-and y-range tail}
}
\value{
plot
}
\description{
Plot Diffusion Map
}
\details{
Plot Diffusion Map or more exactly jump distance map as in Fig 2c of Gasser et al. 2020 doi:10.1017/S0033583519000155
}
\examples{
xmlPath <- "/Users/jgodet/Seafile/MaBibliotheque/Code/TrackMate/nmeth.2808-sv1.xml"
data <- readTrackMateXML(XMLpath = xmlPath)
data$jump<-jump(data, spaceRes=1)
plotDiffusionMap(data = data,traceNb = c(12),nbPix = 5 )
plotDiffusionMap(data = data,nbPix = 5 )
}
\author{
JuG
}
