% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMSD.r
\name{plotMSD}
\alias{plotMSD}
\title{Plot MSD}
\usage{
plotMSD(
  msdData,
  deltaT,
  fitMSD = TRUE,
  printMSDfit = TRUE,
  npoint4fit = 4,
  ...
)
}
\arguments{
\item{msdData}{MSD data (list) (output of calcMSD function)}

\item{deltaT}{time elapsed between two consecutive frame}

\item{fitMSD}{boolean}

\item{printMSDfit}{boolean}

\item{npoint4fit}{number of points to use for MSD fitting}
}
\value{

}
\description{
Plot MSD
}
\examples{


}
\author{
JuG
}
