% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R,
%   R/accessor-methods.R
\docType{methods}
\name{position<-}
\alias{position<-}
\alias{position}
\alias{position,PlotFeatureSet-method}
\alias{position,ggplot-method}
\alias{position,trellis-method}
\alias{position,gTree-method}
\alias{position<-,GGplotFeatureSet-method}
\alias{position,GGplotFeatureSet-method}
\title{Accessors for the position of GGplotFeatureSet objects}
\usage{
position(object) <- value

position(object)

\S4method{position}{PlotFeatureSet}(object)

\S4method{position}{ggplot}(object)

\S4method{position}{trellis}(object)

\S4method{position}{gTree}(object)

\S4method{position}{GGplotFeatureSet}(object) <- value

\S4method{position}{GGplotFeatureSet}(object)
}
\arguments{
\item{object}{An object of class GGplotFeatureSet. For accession, also plot objects of class ggplot.}

\item{value}{A named list of positioning information in each layer of the plot.}
}
\value{
A modified object of class GGplotFeatureSet.
}
\description{
Get or set the position associated with a GGplotFeatureSet object. This should not be called directly.
}
\examples{
library(ggplot2)
plt = qplot(mtcars$mpg, mtcars$cyl)
fs = makeFeatureSet(plt)
position(fs)
}
