% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_progress.R
\name{predict_changes}
\alias{predict_changes}
\title{Predict Changes in a Variable Over Time}
\usage{
predict_changes(
  data,
  min = NULL,
  max = NULL,
  floor = 0,
  ceiling = 100,
  granularity = 0.1,
  lambdas = NULL,
  best = "high",
  speed = FALSE,
  percentiles = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame or list representing the input data model to be used for the prediction}

\item{min}{Optional. Minimum value of `initialvalue` to predict. If `NULL`,
it is set to `floor` (if provided) or `min(data_model$initialvalue)`.}

\item{max}{Optional. Maximum value of `initialvalue` to predict. If `NULL`,
it is set to `ceiling` (if provided) or `max(data_model$initialvalue)`.}

\item{floor}{Numeric or `NULL`. Lower bound for the final level (`initialvalue + change`).
If `NULL`, predictions are unrestricted on the lower end.}

\item{ceiling}{Numeric or `NULL`. Upper bound for the final level.
If `NULL`, predictions are unrestricted on the upper end.}

\item{granularity}{Numeric. Step size for the prediction sequence. Default is `0.1`.}

\item{lambdas}{Optional. A vector of lambda values used in the penalized spline.
Defaults to `0.1 * 1.148^(0:50)`.}

\item{best}{Character string, either \code{"high"} (default) or \code{"low"}, indicating
whether higher or lower values of the initial variable are considered better.}

\item{speed}{Logical. If TRUE, calculate speed of progress scores. For indicators without overall progress, the speed of progress cannot be computed
Default is FALSE}

\item{percentiles}{Logical. If TRUE, calculate percentile scores. Default is TRUE}
}
\value{
A list containing prediction results. The content depends on the selected method:
\describe{
  \item{`predictions_speed`}{Predicted changes using the `"speed"` method (if selected).}
  \item{`path_speed`}{The path or trajectory of predictions based on the `"speed"` method (if selected).}
  \item{`predictions_pctls`}{Predicted changes using the `"percentiles"` method (if selected).}
}
}
\description{
This function calculates projected changes in a variable based on its initial value.
Two methods are available: `"speed"` and `"percentiles"`.
}
\details{
- The `"speed"` method XX TO COMPLETE
- The `"percentiles"` method xx TO COMPLETE
}
\seealso{
[predict_speed()], [get_speed_path()], [predict_pctls()]
}
