% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_changes.R
\name{predict_changes}
\alias{predict_changes}
\title{Predict Changes in an indicator Over Time}
\usage{
predict_changes(
  data,
  min = NULL,
  max = NULL,
  granularity = 0.1,
  best = "high",
  speed = FALSE,
  percentiles = TRUE,
  sequence_pctl = seq(20, 80, 20),
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame or list representing the input data model to be used for the prediction.
This should be the output of the [`prep_data()`] function, which prepares the data in the required format.}

\item{min}{Optional. Minimum value of `initialvalue` to predict. If `NULL`,
it is set to `floor` (if provided) or `min(data_model$initialvalue)`.}

\item{max}{Optional. Maximum value of `initialvalue` to predict. If `NULL`,
it is set to `ceiling` (if provided) or `max(data_model$initialvalue)`.}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.}

\item{best}{A character string indicating the direction of improvement.
Use \code{"high"} (default) if higher values of the indicator are better,
or \code{"low"} if lower values are better. This affects the direction of the path.}

\item{speed}{Logical. If TRUE, calculate speed of progress scores.
Default is FALSE}

\item{percentiles}{Logical. If TRUE, calculate percentile scores. Default is TRUE}

\item{sequence_pctl}{Percentiles to calculate. The percentile score assigned is determined by the granularity of the percentiles selected here. I.e., if seq(20,80,20) is selected, then we will be able to tell if a country is between 0-20, 20-40, etc.}

\item{verbose}{Logical. If TRUE, prints info messages in console}
}
\value{
A list containing prediction results. The content depends on the selected method:
\describe{
  \item{`changes_speed`}{Predicted changes using the `"speed"` method (if selected).}
  \item{`path_speed`}{The path or trajectory of predictions based on the `"speed"` method (if selected).}
  \item{`changes_pctls`}{Predicted changes using the `"percentiles"` method (if selected).}
}
}
\description{
This function calculates projected changes in an indicator based on its initial value.
Two methods are available: `"speed"` and `"percentiles"`.
}
\details{
- On The `"speed"` method: for indicators without overall progress, the speed of progress cannot be computed.
- The `"percentiles"` method can be computed for all indicators but ideally require rich training data.
}
\seealso{
[predict_speed()], [get_speed_path()], [predict_pctls()]
}
