% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_changes.R
\name{predict_changes}
\alias{predict_changes}
\title{Predict Changes in an indicator Over Time}
\usage{
predict_changes(
  data,
  min = NULL,
  max = NULL,
  granularity = 0.1,
  best = "high",
  speed = FALSE,
  percentiles = TRUE,
  sequence_pctl = NULL,
  support = 1,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame or list representing the input data model to be used for the prediction.
This should be the output of the [`prep_data()`] function, which prepares the data in the required format.}

\item{min}{Optional. Minimum value of `initialvalue` to predict.}

\item{max}{Optional. Maximum value of `initialvalue` to predict.}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.
If `NULL`, predictions are unrestricted on the upper end.}

\item{best}{A character string indicating whether higher or lower values of thei indicator are better. Use "high" if higher values of the indicator reflect better outcomes, or "low" if lower values are better.}

\item{speed}{Logical. If TRUE, calculate speed of progress scores. Default is FALSE.}

\item{percentiles}{Logical. If TRUE, calculate percentile scores. Default is TRUE}

\item{sequence_pctl}{Numeric vector of percentile paths to calculate.
The percentile score is determined by the granularity of the percentiles chosen here. For example, if seq(20,80,20) is chosen, then a country’s progress will fall in the 0th-20th percentile, 20th-40th percentile etc.
This argument is only relevant if `percentiles=TRUE`.}

\item{verbose}{Logical. If TRUE, prints info messages in console}
}
\value{
A list containing prediction results. The content depends on the selected method:
\describe{
  \item{`changes_speed`}{Predicted changes using the `"speed"` method (if selected).}
  \item{`path_speed`}{The path or trajectory of predictions based on the `"speed"` method (if selected).}
  \item{`changes_pctls`}{Predicted changes using the `"percentiles"` method (if selected).}
}
}
\description{
This function calculates predicted changes in an indicator based on its initial value. Two methods are available: `speed` and `percentiles`.
}
\details{
- On The `"speed"` method: for indicators without overall progress, the speed of progress cannot be computed.
- The `"percentiles"` method can be computed for all indicators but ideally require rich training data.
}
\seealso{
[predict_speed()], [get_speed_path()], [predict_pctls()]
}
