% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_changes.R
\name{predict_pctls}
\alias{predict_pctls}
\title{Predict percentiles}
\usage{
predict_pctls(
  data_model,
  min = NULL,
  max = NULL,
  granularity = 0.1,
  sequence_pctl = seq(20, 80, 20),
  verbose = TRUE
)
}
\arguments{
\item{data_model}{A data frame with the training data. Obtained through `prep_data()`. Must contain:
- `initialvalue`: numeric, the initial level of the indicator.
- `change`: numeric, the change in the indicator.
- `fold_id`: numeric or factor, used for cross-validation.}

\item{min}{Optional. Minimum value of `initialvalue` to predict.}

\item{max}{Optional. Maximum value of `initialvalue` to predict.}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.
If `NULL`, predictions are unrestricted on the upper end.}

\item{sequence_pctl}{Numeric vector of percentile paths to calculate.
The percentile score is determined by the granularity of the percentiles chosen here. For example, if seq(20,80,20) is chosen, then a country’s progress will fall in the 0th-20th percentile, 20th-40th percentile etc.
This argument is only relevant if `percentiles=TRUE`.}

\item{verbose}{Logical. If TRUE, prints info messages in console}
}
\value{
A data frame with calculated changes by percentiles.
}
\description{
This function calculates the changes as a function of initial values by percentile
}
