% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_paths.R
\name{prep_data_fut}
\alias{prep_data_fut}
\title{Prepare data for future target projections}
\usage{
prep_data_fut(
  data = NULL,
  indicator = "EG.ELC.ACCS.ZS",
  granularity = 0.1,
  code_col = "iso3c",
  year_col = "date",
  support = 1,
  extreme_percentile = getOption("trackr.extreme_pctl"),
  verbose = TRUE
)
}
\arguments{
\item{data}{Optional. A data frame with indicator data.}

\item{indicator}{Character. Name of the column with the values of the indicator data}

\item{granularity}{Numeric. Granularity in outcome variable. All outcome values are rounded to the nearest value reflecting this granularity. For example, if 1 is chosen, then all outcome values are rounded to the nearest integer. Default is 0.1. The lower granularity, the more precise the estimated paths but the longer it takes to run the code.}

\item{code_col}{Character. Name of the column with country codes (or other groupings used to calculate and compare progress)}

\item{year_col}{Character. Name of the column with years.}

\item{support}{Numeric. Reflects minimum number of countries that must have experienced a particular outcome value from startyear_data to endyear_data for a progress score to be calculated. This limits the calculation of scores at extreme outcome values at which there is less support to evaluate what typical progress looks like. Will be overruled by min and max if those are more restrictive. Defaults to 1.}

\item{extreme_percentile}{Numeric (length 1 or 2). Percentile(s) used to define the lower and upper tails of the distribution of initial values. Values in the tails are retained only if there is sufficient country support (`support`). If a single value is provided, it is applied symmetrically to both tails. Defaults to `getOption("trackr.extreme_pctl")`.}

\item{verbose}{Logical. If TRUE print messages in console. Default is TRUE}
}
\value{
A `data.table` with the latest observation per country, including:
\describe{
  \item{code}{Standardized country code (from `code_col`)}
  \item{year}{Year of the latest available observation}
  \item{y}{Value of the indicator}
  \item{y_fut}{Rounded value of `y`, based on the specified granularity}
}
}
\description{
Prepare data for future target projections
}
