% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_paths.R
\name{project_path_speed}
\alias{project_path_speed}
\title{Project speed path}
\usage{
project_path_speed(
  data_his,
  speedseq = c(0.25, 0.5, 1, 2, 4),
  path_speed,
  floor = 0,
  ceiling = 100,
  granularity = 0.1,
  start_year = 2000,
  end_year = 2022,
  min = NULL,
  max = NULL
)
}
\arguments{
\item{data_his}{A `data.table` containing historical values with variables `code`, `year`, and `y_his`.}

\item{speedseq}{numeric vector of speed paths to calculate}

\item{floor}{Numeric or `NULL`.Minimum value of indicator.
If `NULL`, predictions are unrestricted on the lower end.}

\item{ceiling}{Maximum value of indicator (NA if none).
If `NULL`, predictions are unrestricted on the upper end.}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.}

\item{start_year}{The first year to include in the analysis}

\item{end_year}{The last year to include in the analysis}

\item{min}{A numeric value indicating the minimum bound,
typically passed from a previously processed dataset (e.g., `prep_data()` output via `min <- data_model$min`).
This value is stored as an attribute of the returned object for later reference (e.g., in visualizations or simulations).}

\item{max}{A numeric value indicating the maximum bound,
typically passed from a previously processed dataset (e.g., `prep_data()` output via `max <- data_model$max`).
This value is stored as an attribute of the returned object for later reference (e.g., in visualizations or simulations).}
}
\value{
A data frame of projected values under different speeds
}
\description{
Calculates path a country would have taken with various speeds
}
