% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_paths.R
\name{project_path_speed}
\alias{project_path_speed}
\title{Project speed path}
\usage{
project_path_speed(
  data_his,
  speedseq = c(0.25, 0.5, 1, 2, 4),
  path_speed,
  floor,
  ceiling,
  granularity,
  start_year = 2000,
  end_year = 2022,
  min = NULL,
  max = NULL
)
}
\arguments{
\item{data_his}{A `data.table` containing historical values with variables `code`, `year`, and `y_his`.}

\item{speedseq}{numeric vector of speed paths to calculate}

\item{floor}{Numeric or `NULL`.Minimum value of indicator.
If `NULL`, predictions are unrestricted on the lower end.}

\item{ceiling}{Maximum value of indicator (NA if none).
If `NULL`, predictions are unrestricted on the upper end.}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.}

\item{start_year}{The first year to include in the analysis}

\item{end_year}{The last year to include in the analysis}

\item{min}{Optional. Minimum value of `initialvalue` to predict. If `NULL`,
it is set to `floor` (if provided) or `min(data_model$initialvalue)`.}

\item{max}{Optional. Maximum value of `initialvalue` to predict. If `NULL`,
it is set to `ceiling` (if provided) or `max(data_model$initialvalue)`.}
}
\value{
A data frame of projected values under different speeds
}
\description{
Calculates path a country would have taken with various speeds
}
