% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_paths.R
\name{project_path_speed}
\alias{project_path_speed}
\title{Project speed path}
\usage{
project_path_speed(
  data_his,
  sequence_speed = c(0.25, 0.5, 1, 2, 4),
  path_speed,
  granularity = 0.1,
  eval_from = 2000,
  eval_to = 2022,
  min = NULL,
  max = NULL,
  best = "high"
)
}
\arguments{
\item{data_his}{A `data.table` containing historical values with variables `code`, `year`, and `y_his`.}

\item{sequence_speed}{numeric vector of speed paths to calculate}

\item{granularity}{A numeric value indicating the step size used when
computing the trajectory over time. Smaller values produce smoother curves.
Defaults to \code{0.1}.}

\item{eval_from}{The first year to include in the analysis}

\item{eval_to}{The last year to include in the analysis}

\item{min}{A numeric value indicating the minimum bound,}

\item{max}{A numeric value indicating the maximum bound,}

\item{best}{A character string indicating the direction of improvement.
Use \code{"high"} (default) if higher values of the indicator are better,
or \code{"low"} if lower values are better. This affects the direction of the path.}
}
\value{
A data frame of projected values under different speeds
}
\description{
Calculates path a country would have taken with various speeds
}
