% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_paths.R
\name{project_pctls_path}
\alias{project_pctls_path}
\title{Project percentiles paths}
\usage{
project_pctls_path(
  data_his,
  start_year,
  end_year,
  granularity,
  pctlseq,
  predictions_pctl,
  verbose = TRUE
)
}
\arguments{
\item{data_his}{A `data.table` containing historical values with variables `code`, `year`, and `y_his`.}

\item{start_year}{The first year to include in the analysis}

\item{end_year}{The last year to include in the analysis}

\item{granularity}{Numeric. Step size for the prediction sequence. Default is `0.1`.}

\item{pctlseq}{Numeric vector. Sequence of percentiles}

\item{predictions_pctl}{A `data.table` with predicted changes by `initialvalue` and `pctl`.}

\item{verbose}{Logical. Whether to print progress messages.}
}
\value{
A `data.table` with projected values `y_his` by `code`, `year`, and `pctl`.
}
\description{
Simulates year-by-year projected paths of a variable across percentiles,
based on historical values and predicted changes.
}
