% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_paths.R
\name{project_pctls_path}
\alias{project_pctls_path}
\title{Historical percentiles paths}
\usage{
project_pctls_path(
  data_his,
  eval_from = 2000,
  eval_to = 2022,
  granularity = 0.1,
  min = NULL,
  max = NULL,
  sequence_pctl = seq(20, 80, 20),
  changes_pctl,
  verbose = TRUE
)
}
\arguments{
\item{data_his}{A `data.table` containing historical values with variables `code`, `year`, and `y_his`.}

\item{eval_from}{The first year to include in the analysis}

\item{eval_to}{The last year to include in the analysis}

\item{granularity}{Numeric. Granularity in outcome variable. All outcome values are rounded to the nearest value reflecting this granularity. For example, if 1 is chosen, then all outcome values are rounded to the nearest integer. Default is 0.1. The lower granularity, the more precise the estimated paths but the longer it takes to run the code.}

\item{min}{A numeric value indicating the minimum bound,}

\item{max}{A numeric value indicating the maximum bound,}

\item{sequence_pctl}{Numeric vector. Sequence of percentiles}

\item{changes_pctl}{A `data.table` with predicted changes by `initialvalue` and `pctl`.}

\item{verbose}{Logical. Whether to print progress messages.}
}
\value{
A `data.table` with projected values `y_his` by `code`, `year`, and `pctl`.
}
\description{
Simulates year-by-year projected paths of a variable across percentiles, based on historical values and predicted changes
}
