% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_paths.R
\name{project_pctls_path}
\alias{project_pctls_path}
\title{Project percentiles paths}
\usage{
project_pctls_path(
  data_his,
  start_year = 2000,
  end_year = 2022,
  granularity = 0.1,
  min = NULL,
  max = NULL,
  pctlseq = seq(20, 80, 20),
  predictions_pctl,
  verbose = TRUE
)
}
\arguments{
\item{data_his}{A `data.table` containing historical values with variables `code`, `year`, and `y_his`.}

\item{start_year}{The first year to include in the analysis}

\item{end_year}{The last year to include in the analysis}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.}

\item{min}{A numeric value indicating the minimum bound,
typically passed from a previously processed dataset (e.g., `prep_data()` output via `min <- data_model$min`).
This value is stored as an attribute of the returned object for later reference (e.g., in visualizations or simulations).}

\item{max}{A numeric value indicating the maximum bound,
typically passed from a previously processed dataset (e.g., `prep_data()` output via `max <- data_model$max`).
This value is stored as an attribute of the returned object for later reference (e.g., in visualizations or simulations).}

\item{pctlseq}{Numeric vector. Sequence of percentiles}

\item{predictions_pctl}{A `data.table` with predicted changes by `initialvalue` and `pctl`.}

\item{verbose}{Logical. Whether to print progress messages.}
}
\value{
A `data.table` with projected values `y_his` by `code`, `year`, and `pctl`.
}
\description{
Simulates year-by-year projected paths of a variable across percentiles,
based on historical values and predicted changes.
}
