% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrackMateXML.r
\name{readTrackMateXML}
\alias{readTrackMateXML}
\title{read TrackMate XML output files}
\usage{
readTrackMateXML(XMLpath)
}
\arguments{
\item{XMLpath}{path to the xml file}
}
\value{
data frame
}
\description{
read TrackMate XML output files
}
\examples{
xmlPath <- "/Users/jgodet/Seafile/MaBibliotheque/Code/TrackMate/nmeth.2808-sv1.xml"
data <- readTrackMateXML(XMLpath = xmlPath)
jd<-jump(data, spaceRes=1)
hist(jd, breaks=50)
}
\author{
JuG
}
