% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrackMateXML.r
\name{readTrackMateXML}
\alias{readTrackMateXML}
\title{read TrackMate XML output files}
\usage{
readTrackMateXML(XMLpath, multiCore = TRUE)
}
\arguments{
\item{XMLpath}{path to the xml file}

\item{multiCore}{use multi-cores (boolean)}
}
\value{
data frame
}
\description{
read TrackMate XML output files
}
\examples{
xmlPath <- "/Users/jgodet/Seafile/MaBibliotheque/Code/TrackMate/nmeth.2808-sv1.xml"
data <- readTrackMateXML(XMLpath = xmlPath)
data <- readTrackMateXML(XMLpath = xmlPath, multiCore = FALSE)
data$jump<-jump(data, spaceRes=1)
hist(data$jump, breaks=150)
}
\author{
JuG
}
