% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R
\docType{methods}
\name{regDateTime<-}
\alias{regDateTime<-}
\alias{regDateTime}
\alias{regDateTime<-,FeatureSet-method}
\alias{regDateTime,FeatureSet-method}
\title{Accessors for the registration Date/Time of FeatureSet objects}
\usage{
regDateTime(object) <- value

regDateTime(object)

\S4method{regDateTime}{FeatureSet}(object) <- value

\S4method{regDateTime}{FeatureSet}(object)
}
\arguments{
\item{object}{An object of (super)class ObjFeatureSet.}

\item{value}{A POSIXct timestamp representing the moment when the plot was registered.}
}
\value{
A modified object of (super)class ObjFeatureSet.
}
\description{
Get or set the registration time associated with a FeatureSet object. This should not be called directly.
}
\examples{
fs = makeFeatureSet(mtcars)
regDateTime(fs)
}
