% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{select}
\alias{select}
\alias{select_}
\alias{select_.tracks}
\title{Select tables from a tracks object}
\usage{
\method{select_}{tracks}(.data, ..., .dots)
}
\arguments{
\item{.data}{A tracks object}

\item{...}{Comma seperated list of tables to select. Use -table to drop a
table.}

\item{.dots}{Use \code{select_()} to do standard evaluation. See
\code{vignette("nse")} for details}
}
\value{
A tracks object.
}
\description{
This function allows for convenient selection of some of the tables in a
tracks object, usually to manage space or computation times. It will always
retain the tr and meta_data tables, as well as the list of parameters.
}
\details{
You can use dplyr's special functions inside of select calls, such as
starts_with().
}
\examples{
  (guppies_large <- expand_tracks(guppies))
  select(guppies_large, pairs)
  select(guppies_large, -pairs, -animal)
  select(guppies_large, ends_with('l'))
}

