% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R,
%   R/accessor-methods.R
\docType{methods}
\name{statTransform<-}
\alias{statTransform<-}
\alias{statTransform}
\alias{statTransform,PlotFeatureSet-method}
\alias{statTransform,ggplot-method}
\alias{statTransform,trellis-method}
\alias{statTransform,gTree-method}
\alias{statTransform<-,GGplotFeatureSet-method}
\alias{statTransform,GGplotFeatureSet-method}
\title{Accessors for the stat(s) of GGplotFeatureSet objects}
\usage{
statTransform(object) <- value

statTransform(object)

\S4method{statTransform}{PlotFeatureSet}(object)

\S4method{statTransform}{ggplot}(object)

\S4method{statTransform}{trellis}(object)

\S4method{statTransform}{gTree}(object)

\S4method{statTransform}{GGplotFeatureSet}(object) <- value

\S4method{statTransform}{GGplotFeatureSet}(object)
}
\arguments{
\item{object}{An object of class GGplotFeatureSet. For accession, also plot objects of class ggplot.}

\item{value}{A named list of parameters for statistical transforms in each layer of the plot.}
}
\value{
A modified object of class GGplotFeatureSet.
}
\description{
Get or set the stat transform(s)  associated with a GGplotFeatureSet object. This should not be called directly.
}
\examples{
library(ggplot2)
mt = datasets::mtcars
plt = qplot(mt$mpg, mt$cyl)
fs = makeFeatureSet(plt)
statTransform(fs)
}
