% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-dplyr_verbs.R
\name{summarise_sections}
\alias{summarise_sections}
\alias{summarise_sections_}
\alias{summarize_sections}
\alias{summarize_sections_}
\title{Summarize track sections.}
\usage{
summarise_sections(sections, tracks, group_by = NULL, ...)

summarize_sections(sections, tracks, group_by = NULL, ...)

summarize_sections_(sections, tracks, group_by = NULL, ..., .dots)

summarise_sections_(sections, tracks, group_by = NULL, ..., .dots)
}
\arguments{
\item{tracks}{A table with track sections (output from
\code{find_track_sections}).}

\item{group_by}{What factors the result should be grouped by. For example,
\code{~animal} will return a summary for each sequence for each animal.}

\item{...}{Summary statements.}

\item{.dots}{Used to work around non-standard evaluation. See vignette("nse")
for details.}

\item{tracks}{A tracks object.}
}
\value{
A tbl_df.
}
\description{
Similar to \code{filter}, but returns a table of sequences grouped by trial,
with a \code{start}, \code{stop} and \code{length} column. It supports a
tolerance level which allows for combining sequences that are close together
(the default of 1 is zero tolerance).
}
\details{
Any variables used as conditions will be looked up in the \code{tr},
\code{pairs} and \code{group} tables and applied when present. Non-existing
variables will not produce an error.

Seperate conditions on different variables with a \code{,}, not \code{&}.
Combine several conditions on the same variable with \code{&}.

Example use case: If one has selected all sequences of frames where animals
are chasing each other, one can calculate the mean speed for each of the
chase sequences.
}
\section{Functions}{
\itemize{
\item \code{summarize_sections}: Retrieve the timestamps for track section
based on conditions.

\item \code{summarize_sections_}: Retrieve the timestamps for track section
based on conditions.

\item \code{summarise_sections_}: Retrieve the timestamps for track section
based on conditions.
}}

