% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R,
%   R/accessor-methods.R
\docType{methods}
\name{titles<-}
\alias{titles<-}
\alias{titles}
\alias{titles,PlotFeatureSet-method}
\alias{titles,ggplot-method}
\alias{titles,trellis-method}
\alias{titles,gTree-method}
\alias{titles<-,PlotFeatureSet-method}
\alias{titles,PlotFeatureSet-method}
\title{Accessors for the title(s) of FeatureSet objects}
\usage{
titles(object) <- value

titles(object)

\S4method{titles}{PlotFeatureSet}(object)

\S4method{titles}{ggplot}(object)

\S4method{titles}{trellis}(object)

\S4method{titles}{gTree}(object)

\S4method{titles}{PlotFeatureSet}(object) <- value

\S4method{titles}{PlotFeatureSet}(object)
}
\arguments{
\item{object}{An object of (super)class PlotFeatureSet. For accession, also plot objects of class ggplot or trellis.}

\item{value}{A named list of titles of the form list(main = "My title", sub = "My subtitle").}
}
\value{
A modified object of (super)class PlotFeatureSet.
}
\description{
Get or set the titles associated with a FeatureSet object. This should not be called directly.
}
\examples{
library(ggplot2)
plt = qplot(mtcars$mpg, mtcars$cyl)
fs = makeFeatureSet(plt)
titles(fs)
}
