% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{track_progress}
\alias{track_progress}
\title{Track Progress on an Indicator Over Time}
\usage{
track_progress(
  indicator = NULL,
  data = wbstats::wb_data(indicator = indicator, lang = "en", country = "countries_only"),
  code_col = "iso3c",
  year_col = "date",
  startyear_data = 2000,
  endyear_data = 2023,
  eval_from = 2000,
  eval_to = 2022,
  speed = FALSE,
  percentiles = TRUE,
  future = FALSE,
  sequence_pctl = seq(20, 80, 20),
  sequence_speed = c(0.25, 0.5, 1, 2, 4),
  best = NULL,
  target_year = 2030,
  floor = 0,
  ceiling = 100,
  granularity = 0.1,
  min = NULL,
  max = NULL,
  lambdas = 0.1 * 1.148^(0:50),
  verbose = TRUE
)
}
\arguments{
\item{indicator}{Character. Indicator code (e.g., `"EG.ELC.ACCS.ZS"`). Defaults to access to electricity.}

\item{data}{Optional. A data frame with indicator data. If NULL, data is downloaded via `wbstats::wb_data()`.}

\item{code_col}{Character. Name of the column with country codes. Defaults to `"iso3c"`.}

\item{year_col}{Character. Name of the column with years. Defaults to `"date"`.}

\item{startyear_data}{Integer. Minimum year to include in the data. Defaults to 2000.}

\item{endyear_data}{Integer. Maximum year to include in the data. Defaults to 2023.}

\item{eval_from}{First year to include in projections.}

\item{eval_to}{Last year to include in projections.}

\item{speed}{Logical. If TRUE, calculate speed of progress scores.
Default is FALSE}

\item{percentiles}{Logical. If TRUE, calculate percentile scores. Default is TRUE}

\item{sequence_pctl}{Percentiles to calculate. The percentile score assigned is determined by the granularity of the percentiles selected here. I.e., if seq(20,80,20) is selected, then we will be able to tell if a country is between 0-20, 20-40, etc.}

\item{best}{A character string indicating the direction of improvement.
Use \code{"high"} (default) if higher values of the indicator are better,
or \code{"low"} if lower values are better. This affects the direction of the path.}

\item{target_year}{An integer specifying the year to project to. Default is `2030`.}

\item{floor}{Numeric or `NULL`.Minimum value of indicator.
If `NULL`, predictions are unrestricted on the lower end.}

\item{ceiling}{Maximum value of indicator (NA if none).
If `NULL`, predictions are unrestricted on the upper end.}

\item{granularity}{Numeric. Granularity in outcome variable. Default is `0.1`.}

\item{min}{Optional. Minimum value of `initialvalue` to predict. If `NULL`,
it is set to `floor` (if provided) or `min(data_model$initialvalue)`.}

\item{max}{Optional. Maximum value of `initialvalue` to predict. If `NULL`,
it is set to `ceiling` (if provided) or `max(data_model$initialvalue)`.}

\item{lambdas}{Optional. A vector of lambda values, i.e., levels of flexibility that should be tried and evaluated through cross validation
Defaults to `0.1 * 1.148^(0:50)`.}

\item{verbose}{Logical. If TRUE print messages in console. Default is TRUE}
}
\value{
An (invisible) list containing:
\describe{
  \item{data_model}{Cleaned and normalized indicator data.}
  \item{predicted_changes}{Output from `predict_changes()`.}
  \item{data_historical}{Subset of original data used for historical paths.}
  \item{path_historical}{Historical percentile and speed paths.}
  \item{future_path}{(Optional) Projected future paths; `NULL` if `future = FALSE`.}
  \item{scores}{Performance scores based on historical and projected trends.}
}
}
\description{
Calculates progress scores and future targets for a WDI indicator of choice, depending on user selection of speed, percentiles, and future projections.
}
\seealso{
[prep_data()], [predict_changes()], [future_path()], [get_scores()]
}
