% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr-package.R
\docType{package}
\name{trackr-package}
\alias{trackr}
\alias{trackr-package}
\title{trackr: Track Progress and Forecast Targets for WDI Indicators}
\description{
Calculates progress scores and projects future targets for any indicator of interest. Users can choose their preferred methods and parameters to evaluate and project country performance. Depending on the settings, the package returns two to six data frames summarizing percentile-based and speed-based scores, along with historical and projected indicator trajectories under different assumptions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RossanaTat/sdgtrackr}
  \item Report bugs at \url{https://github.com/RossanaTat/sdgtrackr/issues}
}

}
\keyword{internal}
