% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackR.R
\name{trackR}
\alias{trackR}
\title{Simple Video Tracking Software}
\usage{
trackR(tracker = "classic", ...)
}
\arguments{
\item{tracker}{A character string indicating the tracking algorithm to use
(see 'Details').}

\item{...}{Arguments passed to \link[shiny]{runApp}.}
}
\value{
This function does not return anything. A file is saved if the
 tracking is successfully completed.
}
\description{
Simple video tracking software with Shiny-based GUI.
}
\details{
\code{trackR} provides three different tracking algorithms:
 \describe{
    \item{\code{"classic"}:}{This algorithm uses basic background subtraction
      to segment the objects in the image. A background image can be provided
      by the user or can be computed as the mean or median of \code{n} images
      taken at regular intervals throughout the video. \code{"classic"} should
      work well in most lab situations with a constant and homogenous background.}
    \item{\code{"quanta"}:}{This algorithm uses quantized background subtraction
      to segment the objects in the image. In a quantized background image,
      each pixel corresponds to the lower (if the objects are generally darker
      than the background) or upper (if the objects are generally lighter than
      the background) 2.5 percentile of \code{n} images taken at regular
      intervals throughout the video. \code{"quanta"} should work better than
      \code{"classic"} in situations were the background is constant but not
      homogenous (e.g. if the background has different shades).}
    \item{\code{"tracktor"}:}{This algorithm uses adaptive thresholding to
      segment the objects in the image. It is a liberal implementation of the
      \href{https://github.com/vivekhsridhar/tracktor}{tracktor} algorithm
      developed for Python by Vivek Hari Sridhar, Dominique G. Roche and Simon
      Gingins. \code{"tracktor"} should work better than \code{"classic"} and
      \code{"quanta"} in situations were the background is not constant (e.g.
      when the environment changes over time). See \emph{"Sridhar, V. H., Roche,
      D. G., and Gingins, S. (2019). Tracktor: Image-based automated tracking
      of animal movement and behaviour. Methods Ecol. Evol. 10, 691.
      doi:10.1111/2041-210X.13166"} for more details.}
 }
}
\examples{
# Run trackR classic in an external window
\dontrun{trackR("classic")}

# Run trackR quanta in an external window
\dontrun{trackR("quanta")}

# Run trackR tracktor in an external window
\dontrun{trackR("tracktor")}

# Run trackR classic inside RStudio
\dontrun{trackR("classic", launch.browser = rstudioapi::viewer)}

# Run trackR quanta inside RStudio
\dontrun{trackR("quanta", launch.browser = rstudioapi::viewer)}

# Run trackR tracktor inside RStudio
\dontrun{trackR("tracktor", launch.browser = rstudioapi::viewer)}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
