% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackR.R
\name{trackR}
\alias{trackR}
\title{Simple Video Tracking Software}
\usage{
trackR(...)
}
\arguments{
\item{...}{Arguments passed to \link[shiny]{runApp}.}
}
\value{
This function does not return anything. A file is saved if the
 tracking is successfully completed.
}
\description{
Simple video tracking software with Shiny-based GUI.
}
\examples{
# Run trackR in an external window
trackR()

# Run trackR inside RStudio
trackR(launch.browser = rstudioapi::viewer)

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
