% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{TrackrOptions-class}
\alias{TrackrOptions-class}
\alias{TrackrOptions}
\title{Trackr configuration options}
\usage{
TrackrOptions(
  insert_delay = 0,
  img_dir = "./images",
  img_ext = "png",
  backend_opts = list(...),
  ...
)
}
\arguments{
\item{insert_delay}{numeric. delay in seconds between insertions.}

\item{img_dir}{character. Directory to save image files in. This will be normalized via normalizePath}

\item{img_ext}{character. extension to give image files.}

\item{backend_opts}{list. list of options specific to the backend. Currently ignored
by trackr machinery.}

\item{\dots}{additional arguments, which are collected into a list for the default \code{backend_opts} value.}
}
\description{
TrackrOptions objects dictate various behavior
by trackr when interacting with or preparing to interact with backends. See
individual parameters for behaviors controlled in this manner.
}
