% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/PlotWrapper.R
\name{user<-}
\alias{user<-}
\alias{user}
\alias{user<-,FeatureSet-method}
\alias{user,FeatureSet-method}
\title{Accessors for the "user" of FeatureSet objects}
\usage{
user(object) <- value

user(object)

\S4method{user}{FeatureSet}(object) <- value

\S4method{user}{FeatureSet}(object)
}
\arguments{
\item{object}{An object of (super)class ObjFeatureSet.}

\item{value}{A character vector serving as the username associated with the plot object in the database.}
}
\value{
A modified object of (super)class ObjFeatureSet.
}
\description{
Get or set the user associated with a FeatureSet object. This should not be called directly.
}
\examples{
mt = datasets::mtcars
fs = makeFeatureSet(mt)
user(fs)
}
